Simulation Investigation for Empirical Network Analysis

Fits statistical models to longitudinal sets of networks, and to longitudinal sets of networks and behavioral variables. Not only one-mode networks but also two-mode networks and multivariate networks are allowed. The models are stochastic actor-oriented models.

Package "RSienaTest" is the development version, and is distributed through R-Forge, see http://r-forge.r-project.org/R/?group_id=461. Package "RSiena" is the official release.

Details

The main flow of operations of this package is as follows.

Data objects can be created from matrices and vectors using sienaDependent, coCovar, varCovar, coDyadCovar, etc., and finally sienaDataCreate.

Effects are selected using an sienaEffects object, which can be created using getEffects and may be further specified by includeEffects, setEffect, and includeInteraction.

Control of the estimation algorithm requires a sienaAlgorithm object that defines the settings (parameters) of the algorithm, and which can be created by sienaAlgorithmCreate.

Function siena07 is used to fit a model. Function sienaGOF can be used for studying goodness of fit.

A general introduction to the method is available in the tutorial paper Snijders, van de Bunt, and Steglich (2010). Next to the help pages, more detailed help is available in the manual (see below) and a lot of information is at the website (also see below).

Package:	RSiena

Type:	Package

Version:	1.2-25

Date:	2020-04-06

Depends:	R (>= 3.0.0)

Imports:	Matrix

Suggests:	tcltk, network, codetools, lattice, MASS, parallel, xtable, tools, utils

SystemRequirements:	GNU make, tcl/tk 8.5, Tktable

License:	GPL-2

LazyData:	yes

NeedsCompilation:	yes

BuildResaveData:	no

References

Schweinberger, Michael, and Snijders, Tom A.B. (2007). Markov models for digraph panel data: Monte Carlo-based derivative estimation. Computational Statistics and Data Analysis 51, 4465--4483.

Snijders, Tom A.B. (2001). The statistical evaluation of social network dynamics. Sociological Methodology 31, 361-395.

Snijders, Tom A.B. (2017). Stochastic Actor-Oriented Models for Network Dynamics. Annual Review of Statistics and Its Application 4, 343--363.

Snijders, Tom A.B., van de Bunt, Gerhard G., and Steglich, Christian E.G. (2010). Introduction to actor-based models for network dynamics. Social Networks 32, 44--60.

Snijders, Tom A.B., Steglich, Christian E.G., and Schweinberger, Michael (2007). Modeling the co-evolution of networks and behavior. Pp. 41--71 in Longitudinal models in the behavioral and related sciences, edited by Kees van Montfort, Han Oud and Albert Satorra; Lawrence Erlbaum.

Steglich, Christian E.G., Snijders, Tom A.B., and Pearson, Michael A. (2010). Dynamic networks and behavior: Separating selection from influence. Sociological Methodology 40, 329--393.

The manual: http://www.stats.ox.ac.uk/~snijders/siena/RSiena_Manual.pdf

The website: http://www.stats.ox.ac.uk/~snijders/siena/.

See also

Examples

mynet1 <- sienaDependent(array(c(tmp3, tmp4), dim=c(32, 32, 2)))
mydata <- sienaDataCreate(mynet1)
myeff <- getEffects(mydata)
myeff <- includeEffects(myeff, transTrip)

#>   effectName          include fix   test  initialValue parm
#> 1 transitive triplets TRUE    FALSE FALSE          0   0   

myeff

#>   effectName                  include fix   test  initialValue parm
#> 1 basic rate parameter mynet1 TRUE    FALSE FALSE    4.80941   0   
#> 2 outdegree (density)         TRUE    FALSE FALSE   -0.56039   0   
#> 3 reciprocity                 TRUE    FALSE FALSE    0.00000   0   
#> 4 transitive triplets         TRUE    FALSE FALSE    0.00000   0   

myalgorithm <- sienaAlgorithmCreate(nsub=3, n3=200)

#> If you use this algorithm object, siena07 will create an output file Siena.txt .

ans <- siena07(myalgorithm, data=mydata, effects=myeff, batch=TRUE)

#> 
#> Start phase 0 
#> theta: -0.56  0.00  0.00 
#> 
#> Start phase 1 
#> Phase 1 Iteration 1 Progress: 0%
#> Phase 1 Iteration 2 Progress: 0%
#> Phase 1 Iteration 3 Progress: 0%
#> Phase 1 Iteration 4 Progress: 0%
#> Phase 1 Iteration 5 Progress: 0%
#> Phase 1 Iteration 10 Progress: 1%
#> Phase 1 Iteration 15 Progress: 1%
#> Phase 1 Iteration 20 Progress: 2%
#> Phase 1 Iteration 25 Progress: 2%
#> Phase 1 Iteration 30 Progress: 2%
#> Phase 1 Iteration 35 Progress: 3%
#> Phase 1 Iteration 40 Progress: 3%
#> Phase 1 Iteration 45 Progress: 4%
#> Phase 1 Iteration 50 Progress: 4%
#> theta: -0.7478  0.2025  0.0497 
#> 
#> Start phase 2.1
#> Phase 2 Subphase 1 Iteration 1 Progress: 16%
#> Phase 2 Subphase 1 Iteration 2 Progress: 16%
#> theta -0.8852  0.5132  0.0815 
#> ac -0.0872  1.2279  1.1412 
#> Phase 2 Subphase 1 Iteration 3 Progress: 16%
#> Phase 2 Subphase 1 Iteration 4 Progress: 16%
#> theta -1.24  1.14  0.19 
#> ac -0.378  1.236  1.333 
#> Phase 2 Subphase 1 Iteration 5 Progress: 16%
#> Phase 2 Subphase 1 Iteration 6 Progress: 17%
#> theta -1.564  1.361  0.286 
#> ac 0.972 1.288 1.056 
#> Phase 2 Subphase 1 Iteration 7 Progress: 17%
#> Phase 2 Subphase 1 Iteration 8 Progress: 17%
#> theta -1.780  1.639  0.336 
#> ac 0.855 1.047 0.413 
#> Phase 2 Subphase 1 Iteration 9 Progress: 17%
#> Phase 2 Subphase 1 Iteration 10 Progress: 17%
#> theta -1.837  1.664  0.323 
#> ac 0.642 1.103 0.507 
#> theta -1.758  1.376  0.322 
#> ac -0.0577  0.0792 -0.0960 
#> theta: -1.758  1.376  0.322 
#> 
#> Start phase 2.2
#> Phase 2 Subphase 2 Iteration 1 Progress: 34%
#> Phase 2 Subphase 2 Iteration 2 Progress: 34%
#> Phase 2 Subphase 2 Iteration 3 Progress: 34%
#> Phase 2 Subphase 2 Iteration 4 Progress: 34%
#> Phase 2 Subphase 2 Iteration 5 Progress: 35%
#> Phase 2 Subphase 2 Iteration 6 Progress: 35%
#> Phase 2 Subphase 2 Iteration 7 Progress: 35%
#> Phase 2 Subphase 2 Iteration 8 Progress: 35%
#> Phase 2 Subphase 2 Iteration 9 Progress: 35%
#> Phase 2 Subphase 2 Iteration 10 Progress: 35%
#> theta -1.743  1.340  0.316 
#> ac -0.343 -0.207 -0.187 
#> theta: -1.743  1.340  0.316 
#> 
#> Start phase 2.3
#> Phase 2 Subphase 3 Iteration 1 Progress: 55%
#> Phase 2 Subphase 3 Iteration 2 Progress: 55%
#> Phase 2 Subphase 3 Iteration 3 Progress: 55%
#> Phase 2 Subphase 3 Iteration 4 Progress: 56%
#> Phase 2 Subphase 3 Iteration 5 Progress: 56%
#> Phase 2 Subphase 3 Iteration 6 Progress: 56%
#> Phase 2 Subphase 3 Iteration 7 Progress: 56%
#> Phase 2 Subphase 3 Iteration 8 Progress: 56%
#> Phase 2 Subphase 3 Iteration 9 Progress: 56%
#> Phase 2 Subphase 3 Iteration 10 Progress: 56%
#> theta -1.747  1.317  0.321 
#> ac -0.000299 -0.015454 -0.084319 
#> theta: -1.747  1.317  0.321 
#> 
#> Start phase 3 

summary(ans)

#> Estimates, standard errors and convergence t-ratios
#> 
#>                               Estimate   Standard   Convergence 
#>                                            Error      t-ratio   
#> 
#> Rate parameters: 
#>   0       Rate parameter       3.0797  ( 0.4490   )             
#> 
#> Other parameters: 
#>   1. eval outdegree (density) -1.7474  ( 0.2543   )   -0.0303   
#>   2. eval reciprocity          1.3172  ( 0.4122   )   -0.0704   
#>   3. eval transitive triplets  0.3214  ( 0.0948   )   -0.1001   
#> 
#> Overall maximum convergence ratio:    0.1368 
#> 
#> 
#> Total of 784 iteration steps.
#> 
#> Covariance matrix of estimates (correlations below diagonal)
#> 
#>        0.065       -0.045       -0.016
#>       -0.425        0.170       -0.005
#>       -0.683       -0.137        0.009
#> 
#> Derivative matrix of expected statistics X by parameters:
#> 
#>       64.491       29.462      176.719
#>       23.122       21.996       78.738
#>      150.862       76.107      631.380
#> 
#> Covariance matrix of X (correlations below diagonal):
#> 
#>       73.580       37.631      242.588
#>        0.687       40.822      151.843
#>        0.813        0.684     1208.693