Wald and score tests for RSiena results

These functions compute Wald-type and score-type tests for results estimated by siena07.

Wald.RSiena(A, ans)

Multipar.RSiena(ans, ...)

score.Test(ans, test=ans$test)

Arguments

A
A k * p matrix, where p = ans$pp, the number of parameters in ans excluding the basic rate parameters used for conditional estimation.

ans
An object of class sienaFit, resulting from a call to siena07.

...
One or more integer numbers between 1 and p, specifying the tested effects (numbered as in print(ans); if conditional estimation was used, numbered as the 'Other parameters').

test
One or more integer numbers between 1 and p, or a logical vector of length p; these should specify the tested effects (numbered as described for the ...).

Details

The hypothesis tested by Wald.RSiena is \(A\theta = 0\), where \(\theta\) is the parameter estimated in the process leading to ans.

The hypothesis tested by Multipar.RSiena is that all parameters given in \(\ldots\) are 0. This is a special case of the preceding.

The tested effects for score.Test should have been specified in includeEffects or setEffect with fix=TRUE, test=TRUE, i.e., they should not have been estimated. The hypothesis tested by score.Test is that the tested parameters have the value indicated in the effects object used for obtaining ans.

These tests should be carried out only when convergence is adequate (overall maximum convergence ratio less than 0.25 and all t-ratios for convergence less than 0.1 in absolute value).

These functions have their own print method, see print.sienaTest.

Value

An object of class sienaTest, which is a list with elements:

chisquare: The test statistic, assumed to have a chi-squared null distribution.

df: The degrees of freedom.

pvalue: The associated p-value.

onesided: For df=1, the onesided test statistic.

efnames: For Multipar.RSiena and score.Test, the names of the tested effects.

References

See the manual and http://www.stats.ox.ac.uk/~snijders/siena/

See also

Examples

mynet <- sienaDependent(array(c(s501, s502), dim=c(50, 50, 2)))
mydata <- sienaDataCreate(mynet)
myeff <- getEffects(mydata)
myalgorithm <- sienaAlgorithmCreate(nsub=1, n3=40, seed=1777, projname=NULL)

#> If you use this algorithm object, siena07 will create an output file /var/folders/pl/ngry9kn13p7466xwk0_2xyh40000gn/T//RtmpNeONkK/Siena3ebf38249327.txt .
#> This is a temporary file for this R session.

# nsub=1 and n3=40 is used here for having a brief computation,
# not for practice.
myeff <- includeEffects(myeff, transTrip, transTies)

#>   effectName          include fix   test  initialValue parm
#> 1 transitive triplets TRUE    FALSE FALSE          0   0   
#> 2 transitive ties     TRUE    FALSE FALSE          0   0   

myeff <- includeEffects(myeff, outAct, outPop, fix=TRUE, test=TRUE)

#>   effectName             include fix  test initialValue parm
#> 1 outdegree - popularity TRUE    TRUE TRUE          0   0   
#> 2 outdegree - activity   TRUE    TRUE TRUE          0   0   

(ans <- siena07(myalgorithm, data=mydata, effects=myeff, batch=TRUE))

#> 
#> Start phase 0 
#> theta: -1.49  0.00  0.00  0.00  0.00  0.00 
#> 
#> Start phase 1 
#> Phase 1 Iteration 1 Progress: 0%
#> Phase 1 Iteration 2 Progress: 0%
#> Phase 1 Iteration 3 Progress: 0%
#> Phase 1 Iteration 4 Progress: 1%
#> Phase 1 Iteration 5 Progress: 1%
#> Phase 1 Iteration 10 Progress: 2%
#> Phase 1 Iteration 15 Progress: 2%
#> Phase 1 Iteration 20 Progress: 3%
#> Phase 1 Iteration 25 Progress: 4%
#> Phase 1 Iteration 30 Progress: 5%
#> Phase 1 Iteration 35 Progress: 6%
#> Phase 1 Iteration 40 Progress: 6%
#> Phase 1 Iteration 45 Progress: 7%
#> Phase 1 Iteration 50 Progress: 8%
#> theta: -1.750  0.437  0.650 -0.477  0.000  0.000 
#> 
#> Start phase 2.1
#> Phase 2 Subphase 1 Iteration 1 Progress: 56%
#> Phase 2 Subphase 1 Iteration 2 Progress: 57%
#> theta -1.874  0.671  0.830 -0.362  0.000  0.000 
#> ac  0.00378  3.79486 -1.06273 -3.09687 -0.05719  0.12148 
#> Phase 2 Subphase 1 Iteration 3 Progress: 57%
#> Phase 2 Subphase 1 Iteration 4 Progress: 57%
#> theta -2.1545  1.2009  0.6376  0.0483  0.0000  0.0000 
#> ac  0.235  2.246 -1.056 -2.625  0.220  0.408 
#> Phase 2 Subphase 1 Iteration 5 Progress: 57%
#> Phase 2 Subphase 1 Iteration 6 Progress: 57%
#> theta -2.454  1.711  0.230  0.637  0.000  0.000 
#> ac  0.225  1.826 -1.110 -2.672  0.246  0.472 
#> Phase 2 Subphase 1 Iteration 7 Progress: 57%
#> Phase 2 Subphase 1 Iteration 8 Progress: 58%
#> theta -2.621  1.956  0.115  0.669  0.000  0.000 
#> ac -0.2214  0.0523 -1.0209 -1.1194 -0.0999  0.2100 
#> Phase 2 Subphase 1 Iteration 9 Progress: 58%
#> Phase 2 Subphase 1 Iteration 10 Progress: 58%
#> theta -2.6549  2.1076  0.0503  0.6892  0.0000  0.0000 
#> ac -0.298 -0.225 -1.081 -1.250 -0.230  0.104 
#> theta -2.620  2.096  0.307  0.540  0.000  0.000 
#> ac -0.4355 -0.5384 -0.7333 -0.9225 -0.1641 -0.0898 
#> theta: -2.620  2.096  0.307  0.540  0.000  0.000 
#> 
#> Start phase 3 

#> Estimates, standard errors and convergence t-ratios
#> 
#>                                  Estimate   Standard   Convergence 
#>                                               Error      t-ratio   
#> 
#> Rate parameters: 
#>   0       Rate parameter          6.6977  ( 1.0182   )             
#> 
#> Other parameters: 
#>   1. eval outdegree (density)    -2.6200  ( 0.1514   )   -0.4627   
#>   2. eval reciprocity             2.0958  ( 0.2976   )   -0.5244   
#>   3. eval transitive triplets     0.3066  ( 0.2922   )   -0.8100   
#>   4. eval transitive ties         0.5397  ( 0.4471   )   -0.6856   
#>   5. eval outdegree - popularity  0.0000  (     NA   )    0.1821   
#>   6. eval outdegree - activity    0.0000  (     NA   )    0.1142   
#> 
#> Overall maximum convergence ratio:    0.9363 
#> 
#> 
#> Score test for 2 parameters:
#> chi-squared = 6.33, p = 0.0422.
#> 
#> Total of 122 iteration steps.
#> 

A <- matrix(0, 2, 6)
A[1, 3] <- 1
A[2, 4] <- 1
wa <- Wald.RSiena(A, ans)
wa

#> chi-squared = 25.49, d.f. = 2;  p < 0.001. 

# A shortcut for the above is:
Multipar.RSiena(ans, 3, 4)

#> Tested effects:
#>  mynet: transitive triplets 
#>  mynet: transitive ties 
#> chi-squared = 25.49, d.f. = 2;  p < 0.001. 

# The following two are equivalent:
sct <- score.Test(ans, c(FALSE, FALSE, FALSE, FALSE, FALSE, TRUE))
sct <- score.Test(ans,6)
print(sct)

#> Tested effects:
#>  mynet: outdegree - activity 
#> chi-squared = 3.56, d.f. = 1; one-sided Z = -1.89;  p = 0.059. 

# Getting all 1-df score tests separately:
# (More identifying information for the effects may be added as necessary)
for (i in which(ans$test)){
   sct <- score.Test(ans,i)
   cat(ans$requestedEffects$effectName[i], '\n')
   print(sct)}

#> outdegree - popularity 
#> Tested effects:
#>  mynet: outdegree - popularity 
#> chi-squared = 6.10, d.f. = 1; one-sided Z = -2.47;  p = 0.014. 
#> outdegree - activity 
#> Tested effects:
#>  mynet: outdegree - activity 
#> chi-squared = 3.56, d.f. = 1; one-sided Z = -1.89;  p = 0.059.