Network data: excerpt from "Dutch Social Behavior Data Set" of Chris Baerveldt.

Matrices N3401, N3403, N3404, N3406, and HN3401, HN3403, HN3404, HN3406 are two waves of networks for four schools (numbered 1, 3, 4, 6): there is a tie from pupil i to pupil j if i says that he/she receives and/or gives emotional support from/to pupil j. The data are part of a larger data set (see source below) and were collected under the direction of Chris Baerveldt.

Format

Adjacency matrices for the network at two time points. The matrices with name N... are the first wave, those with name HN... are the second wave.

Source

http://www.stats.ox.ac.uk/~snijders/siena/CB_data.zip

References

Houtzager, B. & Baerveldt, C. (1999). Just like Normal. A Social Network Study of the Relation between Petty Crime and the Intimacy of Adolescent Friendships. Social Behavior and Personality 27(2), 177-192.

Snijders, Tom A.B, and Baerveldt, Chris (2003). A Multilevel Network Study of the Effects of Delinquent Behavior on Friendship Evolution. Journal of Mathematical Sociology 27, 123-151.

See http://www.stats.ox.ac.uk/~snijders/siena/BaerveldtData.html

Examples

mynet <- sienaDependent(array(c(N3401, HN3401), dim=c(45, 45, 2)))
mydata <- sienaDataCreate(mynet)