Function to create user-specified interactions for a Siena model.

This function allows the user to include or exclude an interaction effect in a Siena effects object.

includeInteraction(myeff, ..., include = TRUE, name = myeff$name[1],
    type = "eval", interaction1 = rep("", 3), interaction2 = rep("", 3),
    fix = FALSE, test = FALSE, parameter = NULL, random = FALSE,
    character = FALSE, verbose = TRUE)

Arguments

myeff
a Siena effects object as created by getEffects

...
2 or 3 short names to identify the effects which should be interacted.

include
Boolean. default TRUE, but can be switched to FALSE to turn off an interaction.

name
Name of dependent variable (network or behavior) for which interactions are being defined. Defaults to the first in the effects object.

type
Type of effects to be interacted.

interaction1
Vector of Siena objects where needed to completely identify the effect e.g. covariate name or behavior variable name. Trailing blanks may be omitted.

interaction2
Vector of siena objects where needed to completely identify the effect e.g. covariate name or behavior variable name. Trailing blanks may be omitted.

fix
Boolean. Are the effects to be fixed at the value stored in myeff$initialValue or not.

test
Boolean. Are the effects to be tested or not (requires fix).

parameter
Value of internal effect parameter of this interaction effect. If NULL, no change is made.

random
For specifying that the interaction effect will vary randomly; not relevant for RSiena at this moment. Boolean required. Default FALSE.

character
Boolean: are the effect names character strings or not.

verbose
Boolean: should the print of altered effects be produced.

Details

The details provided should uniquely identify up to three effects. If so, an interaction effect will be created and included or not in the model.
Whether interactions between two or three given effects can be created depends on their interactionType (which can be, for dependent network variables, empty, ego, or dyadic; and for dependent behavioral variables, empty or OK). Consult the section on Interaction Effects in the manual for this. The interactionType is shown in the list of effects obtained from the function effectsDocumentation.
The short names must not be set between quotes, unless you use character=TRUE.
From the point of view of model building it is usually advisable, when including an interaction effect in a model, also to include the corresponding main effects. This is however not enforced by includeInteraction().

Interaction effects are constructed from effects with shortName unspInt (for networks) and behUnspInt (for behavior) by specifying their elements effect1 and effect2, and possibly effect3.

The number of possible user-specified interaction effects is limited by the parameters nintn (for dependent network variables) and behNintn (for dependent behavior variables) in the call of getEffects, which determine the numbers of effects with shortNames unspInt and behUnspInt.

The input names interaction1 and interaction2 do not themselves refer to created interactions, but to dependence of the base effects on other variables in the data set. They are used to completely identify the effects.

To set attributes of interaction effects in the effects object, function setEffect can also be used with short name unspInt or behUnspInt, and further using parameters effect1 and effect2, and possibly effect3.

Further information about Siena effects objects is given in the help page for getEffects.

A list of all effects in a given effects object (e.g., myeff), including their names of dependent variables, effect names, short names, and values of interaction1 and interaction2 (if any), is obtained by executing effectsDocumentation(myeff).

Value

An updated version of the input effects object; if include, containing the interaction effect between "effect1" and "effect2" and possibly "effect3"; if not, without this interaction effect. The shortName of the interaction effect is "unspInt" for network effects and "behUnspInt" for behavior effects.
If verbose=TRUE, details of the fields altered will be printed.

References

See http://www.stats.ox.ac.uk/~snijders/siena/

See also

Examples

mynet <- sienaDependent(array(c(s501, s502, s503), dim=c(50, 50, 3)))
alc <- varCovar(s50a)
mydata <- sienaDataCreate(mynet, alc)
myeff <- getEffects(mydata)
myeff <- includeEffects(myeff, transTrip)

#>   effectName          include fix   test  initialValue parm
#> 1 transitive triplets TRUE    FALSE FALSE          0   0   

myeff <- includeInteraction(myeff, recip, outAct)

#>   effectName                     include fix   test  initialValue parm effect1
#> 1 unspecified interaction effect TRUE    FALSE FALSE          0   0    12     
#>   effect2
#> 1 95     

myeff <- includeEffects(myeff, egoX, altX, simX, interaction1="alc")

#>   effectName     include fix   test  initialValue parm
#> 1 alc alter      TRUE    FALSE FALSE          0   0   
#> 2 alc ego        TRUE    FALSE FALSE          0   0   
#> 3 alc similarity TRUE    FALSE FALSE          0   0   

myeff <- includeInteraction(myeff, recip, simX, interaction1=c("", "alc"))

#>   effectName                     include fix   test  initialValue parm effect1
#> 1 unspecified interaction effect TRUE    FALSE FALSE          0   0    12     
#>   effect2
#> 1 218    

myeff

#>    effectName                     include fix   test  initialValue parm effect1
#> 1  constant mynet rate (period 1) TRUE    FALSE FALSE    4.69604   0     0     
#> 2  constant mynet rate (period 2) TRUE    FALSE FALSE    4.32885   0     0     
#> 3  outdegree (density)            TRUE    FALSE FALSE   -1.46770   0     0     
#> 4  reciprocity                    TRUE    FALSE FALSE    0.00000   0     0     
#> 5  transitive triplets            TRUE    FALSE FALSE    0.00000   0     0     
#> 6  alc alter                      TRUE    FALSE FALSE    0.00000   0     0     
#> 7  alc ego                        TRUE    FALSE FALSE    0.00000   0     0     
#> 8  alc similarity                 TRUE    FALSE FALSE    0.00000   0     0     
#> 9  unspecified interaction effect TRUE    FALSE FALSE    0.00000   0    12     
#> 10 unspecified interaction effect TRUE    FALSE FALSE    0.00000   0    12     
#>    effect2
#> 1    0    
#> 2    0    
#> 3    0    
#> 4    0    
#> 5    0    
#> 6    0    
#> 7    0    
#> 8    0    
#> 9   95    
#> 10 218    

# How to set the effect parameter of an interaction:
myeff <- getEffects(mydata)
myeff <- setEffect(myeff, gwespFF, parameter=69)

#>   effectName            include fix   test  initialValue parm
#> 1 GWESP I -> K -> J (#) TRUE    FALSE FALSE          0   69  

myeff <- includeInteraction(myeff, recip, gwespFF, parameter=69)

#>   effectName                     include fix   test  initialValue parm effect1
#> 1 unspecified interaction effect TRUE    FALSE FALSE          0   69   12     
#>   effect2
#> 1 46     

myeff

#>   effectName                     include fix   test  initialValue parm effect1
#> 1 constant mynet rate (period 1) TRUE    FALSE FALSE    4.69604    0    0     
#> 2 constant mynet rate (period 2) TRUE    FALSE FALSE    4.32885    0    0     
#> 3 outdegree (density)            TRUE    FALSE FALSE   -1.46770    0    0     
#> 4 reciprocity                    TRUE    FALSE FALSE    0.00000    0    0     
#> 5 GWESP I -> K -> J (#)          TRUE    FALSE FALSE    0.00000   69    0     
#> 6 unspecified interaction effect TRUE    FALSE FALSE    0.00000   69   12     
#>   effect2
#> 1  0     
#> 2  0     
#> 3  0     
#> 4  0     
#> 5  0     
#> 6 46