Functions to plot assessment of time heterogeneity of parameters

Plot method for sienaTimeTest objects.

# S3 method for sienaTimeTest
plot(x, pairwise=FALSE, effects,
         scale=.2, plevels=c(.1, .05, .025), ...)

Arguments

x
A sienaTimeTest object returned by sienaTimeTest.

pairwise
A Boolean value corresponding to whether the user would like a pairwise plot of the simulated statistics to assess correlation among the effects (pairwise=TRUE), or a plot of the estimates across waves in order to assess graphically the results of the score type test.

effects
A vector of integers corresponding to the indices given in the sienaTimeTest output for effects which are to be plotted.

scale
A positive number corresponding to the number of standard deviations on one step estimates to use for computing the maximum and minimum of the plotting range. We recommend experimenting with this number when the y-axes of the plots are not satisfactory. Smaller numbers shrink the axes.

plevels
A list of three decimals indicating the gradients at which to draw the confidence interval bars.

...
For extra arguments. The Lattice parameter layout can be used to control the layout of the graphs.

Details

The pairwise=TRUE plot may be used to assess whether effects are highly correlated. This information may be important when considering forward-model selection, since highly correlated effects may have highly correlated one-step estimates, particularly since the individual score type tests are not orthogonalized against the scores and deviations of yet-unestimated dummies. For example, reciprocity and outdegree may have highly correlated statistics as indicated by a strong, positive correlation coefficient. When considering whether to include dummy terms, it may be a good idea to include, e.g., outdegree, estimate the parameter, and see whether reciprocity dummies remain significant after method of moments estimation of the updated model--as opposed to including both outdegree and reciprocity.

The pairwise=FALSE plot displays the most of the information garnered from sienaTimeTest in a graphical fashion. For each effect, the method of moments parameter estimate for the base period (i.e. wave 1) is given as a blue, horizontal reference line. One step estimates are given for all of the parameters by dots at each wave. The dots are colored black if the parameter has been included in the model already (i.e. has been estimated via method of moments), or red if they have not been included. Confidence intervals are given based on pivots given at pvalues. Evidence of time heterogeneity is suggested by points with confidence intervals not overlapping with the base period.

Value

None

References

See http://www.stats.ox.ac.uk/~snijders/siena/ for general information on RSiena.

See also

Examples

if (FALSE) {
myalgorithm <- sienaAlgorithmCreate(nsub=4, n3=500)
mynet1 <- sienaDependent(array(c(s501, s502, s503, s501, s503, s502), dim=c(50, 50, 6)))
mydata <- sienaDataCreate(mynet1)
myeff <- getEffects(mydata)
myeff <- includeEffects(myeff, transTrip, balance)
myeff <- includeTimeDummy(myeff, recip, timeDummy="2,3,5")
myeff <- includeTimeDummy(myeff, balance, timeDummy="4")
myeff <- includeTimeDummy(myeff, density, timeDummy="all")
ansp <- siena07(myalgorithm, data=mydata, effects=myeff, batch=TRUE)
ttp <- sienaTimeTest(ansp)

## Pairwise plots show
plot(ttp, pairwise=TRUE)

## Time test plots show
plot(ttp, effects=1:3) ## default layout
plot(ttp, effects=1:3, layout=c(3,1))
}