Print methods for Siena effects objects

Print the major columns of the effects object. Or all, with any non atomic columns listed separately.

# S3 method for sienaEffects
print(x, fileName = NULL, includeOnly=TRUE,
expandDummies = FALSE, includeRandoms = FALSE, dropRates=FALSE, ...)
# S3 method for sienaEffects
summary(object, fileName = NULL,
includeOnly=TRUE, expandDummies = FALSE, ...)
# S3 method for summary.sienaEffects
print(x, fileName = NULL, ...)

Arguments

object
An object of class sienaEffects.

x
An object of class sienaEffects or summary.sienaEffects as appropriate.

fileName
Character string denoting file name if file output desired.

includeOnly
Boolean. If TRUE, only effects with the include flag TRUE will be printed.

expandDummies
Interpret the timeDummy column and show any effects which would be added by sienaTimeFix.

includeRandoms
Boolean. If TRUE, also the randomEffects column will be printed.

dropRates
Boolean. If TRUE, do not print the rows for basic rate effects.

...
For extra arguments (none used at present).

Value

The function print.sienaEffects prints details of the main columns of the selected rows of the effects object.

The function summary.sienaEffects checks the rows for valid printing via print.data.frame and excludes any that will fail. The OK columns are printed first, followed by any others.

Output from either can be directed to a file by using the argument filename.

References

See http://www.stats.ox.ac.uk/~snijders/siena/

See also

Examples

mynet1 <- sienaDependent(array(c(s501, s502, s503), dim=c(50, 50, 3)))
mybeh <- sienaDependent(s50a, type="behavior")
mycovar <- coCovar(rnorm(50))
mydyadcovar <- coDyadCovar(matrix(as.numeric(rnorm(2500) > 2), nrow=50))
mydata <- sienaDataCreate(mynet1, mybeh, mycovar, mydyadcovar)
myeff <- getEffects(mydata)
myeff

#>   name   effectName                      include fix   test  initialValue parm
#> 1 mynet1 constant mynet1 rate (period 1) TRUE    FALSE FALSE    4.69604   0   
#> 2 mynet1 constant mynet1 rate (period 2) TRUE    FALSE FALSE    4.32885   0   
#> 3 mynet1 outdegree (density)             TRUE    FALSE FALSE   -1.46770   0   
#> 4 mynet1 reciprocity                     TRUE    FALSE FALSE    0.00000   0   
#> 5 mybeh  rate mybeh (period 1)           TRUE    FALSE FALSE    0.70571   0   
#> 6 mybeh  rate mybeh (period 2)           TRUE    FALSE FALSE    0.84939   0   
#> 7 mybeh  mybeh linear shape              TRUE    FALSE FALSE    0.32237   0   
#> 8 mybeh  mybeh quadratic shape           TRUE    FALSE FALSE    0.00000   0   

summary(myeff)

#>                     name                      effectName
#> mynet1.rate.1     mynet1 constant mynet1 rate (period 1)
#> mynet1.rate.2     mynet1 constant mynet1 rate (period 2)
#> mynet1.obj.eval.1 mynet1             outdegree (density)
#> mynet1.obj.eval.2 mynet1                     reciprocity
#> mybeh.rate.1       mybeh           rate mybeh (period 1)
#> mybeh.rate.2       mybeh           rate mybeh (period 2)
#> mybeh.obj.eval.1   mybeh              mybeh linear shape
#> mybeh.obj.eval.2   mybeh           mybeh quadratic shape
#>                                                       functionName shortName
#> mynet1.rate.1                 Amount of network change in period 1      Rate
#> mynet1.rate.2                 Amount of network change in period 2      Rate
#> mynet1.obj.eval.1                                   Number of ties   density
#> mynet1.obj.eval.2                      Number of reciprocated ties     recip
#> mybeh.rate.1      Amount of behavioral change in period 1 on mybeh      Rate
#> mybeh.rate.2      Amount of behavioral change in period 2 on mybeh      Rate
#> mybeh.obj.eval.1                                mybeh centered sum    linear
#> mybeh.obj.eval.2                        mybeh sum of cent. squares      quad
#>                   interaction1 interaction2 type basicRate include
#> mynet1.rate.1                               rate      TRUE    TRUE
#> mynet1.rate.2                               rate      TRUE    TRUE
#> mynet1.obj.eval.1                           eval     FALSE    TRUE
#> mynet1.obj.eval.2                           eval     FALSE    TRUE
#> mybeh.rate.1                                rate      TRUE    TRUE
#> mybeh.rate.2                                rate      TRUE    TRUE
#> mybeh.obj.eval.1                            eval     FALSE    TRUE
#> mybeh.obj.eval.2                            eval     FALSE    TRUE
#>                   randomEffects   fix  test timeDummy initialValue parm
#> mynet1.rate.1             FALSE FALSE FALSE         ,    4.6960424    0
#> mynet1.rate.2             FALSE FALSE FALSE         ,    4.3288454    0
#> mynet1.obj.eval.1          TRUE FALSE FALSE         ,   -1.4677046    0
#> mynet1.obj.eval.2         FALSE FALSE FALSE         ,    0.0000000    0
#> mybeh.rate.1              FALSE FALSE FALSE         ,    0.7057143    0
#> mybeh.rate.2              FALSE FALSE FALSE         ,    0.8493878    0
#> mybeh.obj.eval.1          FALSE FALSE FALSE         ,    0.3223690    0
#> mybeh.obj.eval.2          FALSE FALSE FALSE         ,    0.0000000    0
#>                   functionType period rateType untrimmedValue effect1 effect2
#> mynet1.rate.1             rate      1     <NA>       0.000000       0       0
#> mynet1.rate.2             rate      2     <NA>       0.000000       0       0
#> mynet1.obj.eval.1    objective   <NA>     <NA>      -1.467705       0       0
#> mynet1.obj.eval.2    objective   <NA>     <NA>       0.000000       0       0
#> mybeh.rate.1              rate      1     <NA>       0.000000       0       0
#> mybeh.rate.2              rate      2     <NA>       0.000000       0       0
#> mybeh.obj.eval.1     objective   <NA>     <NA>       0.322369       0       0
#> mybeh.obj.eval.2     objective   <NA>     <NA>       0.000000       0       0
#>                   effect3 interactionType local setting effectFn statisticFn
#> mynet1.rate.1           0                  TRUE             NULL        NULL
#> mynet1.rate.2           0                  TRUE             NULL        NULL
#> mynet1.obj.eval.1       0          dyadic  TRUE             NULL        NULL
#> mynet1.obj.eval.2       0          dyadic  TRUE             NULL        NULL
#> mybeh.rate.1            0                 FALSE             NULL        NULL
#> mybeh.rate.2            0                 FALSE             NULL        NULL
#> mybeh.obj.eval.1        0              OK FALSE             NULL        NULL
#> mybeh.obj.eval.2        0                 FALSE             NULL        NULL
#>                    netType groupName group effectNumber
#> mynet1.rate.1      oneMode    Group1     1            1
#> mynet1.rate.2      oneMode    Group1     1            2
#> mynet1.obj.eval.1  oneMode    Group1     1           10
#> mynet1.obj.eval.2  oneMode    Group1     1           13
#> mybeh.rate.1      behavior    Group1     1          532
#> mybeh.rate.2      behavior    Group1     1          533
#> mybeh.obj.eval.1  behavior    Group1     1          546
#> mybeh.obj.eval.2  behavior    Group1     1          549