Function to set various columns in an effects object in a Siena model.

This function provides an interface to change various columns of a selected row of a Siena effects object.

setEffect(myeff, shortName, parameter = NULL, fix = FALSE,
test = FALSE, random=FALSE, initialValue = 0, timeDummy = ",", include = TRUE,
name = myeff$name[1], type = "eval", interaction1 = "",
interaction2 = "", effect1=0, effect2=0, effect3=0,
period=1, group=1, character=FALSE, verbose = TRUE)

Arguments

myeff
a Siena effects object as created by getEffects

shortName
A short name (all with or all without quotes) to identify the effect which should be changed.

parameter
Value of internal effect parameter. If NULL, no change is made.

fix
For fixing effects. Boolean required. Default FALSE.

test
For testing effects by score-type tests. Boolean required. Default FALSE.

random
For specifying that effects will vary randomly; not relevant for RSiena at this moment. Boolean required. Default FALSE.

initialValue
Initial value required. Default 0.

timeDummy
string: Comma delimited string of which periods to dummy. Alternatively, use includeTimeDummy.

include
Boolean. default TRUE, but can be switched to FALSE to turn off an effect.

name
Name of dependent variable (network or behavior) for which effects are being modified. Defaults to the first in the effects object.

type
Character string indicating the type of the effect to be changed : currently "rate", "eval", "endow", or "creation". Default "eval".

interaction1
Name of siena object where needed to completely identify the effect e.g. covariate name or behavior variable name.

interaction2
Name of siena object where needed to completely identify the effect e.g. covariate name or behavior variable name.

effect1
Only for shortName=unspInt, which means this is a user-defined interaction effect: effect1 is a natural number indicating the first component of the interaction effect; the number is the one listed when applying print() to myeff.

effect2
Only for shortName=unspInt: second component of interaction effect (see effect1).

effect3
Only for shortName=unspInt: third component of interaction effect, if any (see effect1).

period
Number of period if basic rate. Use numbering within groups.

group
Number of group if basic rate. Only relevant for sienaGroup data sets.

character
Boolean: whether the short name is a character string.

verbose
Boolean: should the print of altered effects be produced.

Details

Recall from the help page for getEffects that a Siena effects object (class sienaEffects or sienaGroupEffects) is a data.frame; the rows in the data frame are the effects for this data set; some of the columns/variables of the data frame are used to identify the effect, other columns/variables define how this effect is used in the estimation.
The arguments shortName, name, type, interaction1, interaction2, effect1, effect2, effect3, period, and group should identify one effect completely. (Not all of them are needed; see getEffects.)
The call of setEffect will set, for this effect, the column elements of the resulting effects object for parameter, fix, test, randomEffects, initialValue, timeDummy, and include to the values requested.
The shortName must not be set between quotes, unless you use character=TRUE.

Value

An object of class sienaEffects or sienaGroupEffects. This will be an updated version of the input effects object, with one row updated. Details of the row altered will be printed, unless verbose=FALSE.

References

See http://www.stats.ox.ac.uk/~snijders/siena/

See also

Examples

mynet <- sienaDependent(array(c(s501, s502, s503), dim=c(50, 50, 3)))
mybeh <- sienaDependent(s50a, type="behavior")
mydata <- sienaDataCreate(mynet, mybeh)
myeff <- getEffects(mydata)
# Set the initial rate parameter for one period:
myeff <- setEffect(myeff, Rate, initialValue=1.5, name="mybeh",
                   type="rate", period=2)

#>   effectName            include fix   test  initialValue parm
#> 1 rate mybeh (period 2) TRUE    FALSE FALSE        1.5   0   

# Specify an effect parameter:
myeff <- setEffect(myeff, outTrunc, parameter=1)

#>   effectName         include fix   test  initialValue parm
#> 1 outdegree-trunc(#) TRUE    FALSE FALSE          0   1   

myeff

#>   name  effectName                     include fix   test  initialValue parm
#> 1 mynet constant mynet rate (period 1) TRUE    FALSE FALSE    4.69604   0   
#> 2 mynet constant mynet rate (period 2) TRUE    FALSE FALSE    4.32885   0   
#> 3 mynet outdegree (density)            TRUE    FALSE FALSE   -1.46770   0   
#> 4 mynet reciprocity                    TRUE    FALSE FALSE    0.00000   0   
#> 5 mynet outdegree-trunc(#)             TRUE    FALSE FALSE    0.00000   1   
#> 6 mybeh rate mybeh (period 1)          TRUE    FALSE FALSE    0.70571   0   
#> 7 mybeh rate mybeh (period 2)          TRUE    FALSE FALSE    1.50000   0   
#> 8 mybeh mybeh linear shape             TRUE    FALSE FALSE    0.32237   0   
#> 9 mybeh mybeh quadratic shape          TRUE    FALSE FALSE    0.00000   0   

# Now request the interaction between reciprocity and alcohol similarity
# to be tested without estimation.
myeff <- includeInteraction(myeff, recip, simX, interaction1=c("", "mybeh"))

#>   effectName                     include fix   test  initialValue parm effect1
#> 1 unspecified interaction effect TRUE    FALSE FALSE          0   0    12     
#>   effect2
#> 1 218    

myeff

#>    name  effectName                     include fix   test  initialValue parm
#> 1  mynet constant mynet rate (period 1) TRUE    FALSE FALSE    4.69604   0   
#> 2  mynet constant mynet rate (period 2) TRUE    FALSE FALSE    4.32885   0   
#> 3  mynet outdegree (density)            TRUE    FALSE FALSE   -1.46770   0   
#> 4  mynet reciprocity                    TRUE    FALSE FALSE    0.00000   0   
#> 5  mynet outdegree-trunc(#)             TRUE    FALSE FALSE    0.00000   1   
#> 6  mynet unspecified interaction effect TRUE    FALSE FALSE    0.00000   0   
#> 7  mybeh rate mybeh (period 1)          TRUE    FALSE FALSE    0.70571   0   
#> 8  mybeh rate mybeh (period 2)          TRUE    FALSE FALSE    1.50000   0   
#> 9  mybeh mybeh linear shape             TRUE    FALSE FALSE    0.32237   0   
#> 10 mybeh mybeh quadratic shape          TRUE    FALSE FALSE    0.00000   0   
#>    effect1 effect2
#> 1   0        0    
#> 2   0        0    
#> 3   0        0    
#> 4   0        0    
#> 5   0        0    
#> 6  12      218    
#> 7   0        0    
#> 8   0        0    
#> 9   0        0    
#> 10  0        0    

# The following code uses the fact that the interaction effect is number 6.
# In practice one would just look up the numbers of effect1 and effect2
# from the response to \code{myeff} and fill them in.
(eff1 <- myeff[myeff$include,]$effect1[6])

#> [1] 12

(eff2 <- myeff[myeff$include,]$effect2[6])

#> [1] 218

myeff <- setEffect(myeff, unspInt, fix=TRUE, test=TRUE,
                                    effect1=eff1, effect2=eff2)

#>   effectName                     include fix  test initialValue parm effect1
#> 1 unspecified interaction effect TRUE    TRUE TRUE          0   0    12     
#>   effect2
#> 1 218    

myeff

#>    name  effectName                     include fix   test  initialValue parm
#> 1  mynet constant mynet rate (period 1) TRUE    FALSE FALSE    4.69604   0   
#> 2  mynet constant mynet rate (period 2) TRUE    FALSE FALSE    4.32885   0   
#> 3  mynet outdegree (density)            TRUE    FALSE FALSE   -1.46770   0   
#> 4  mynet reciprocity                    TRUE    FALSE FALSE    0.00000   0   
#> 5  mynet outdegree-trunc(#)             TRUE    FALSE FALSE    0.00000   1   
#> 6  mynet unspecified interaction effect TRUE    TRUE  TRUE     0.00000   0   
#> 7  mybeh rate mybeh (period 1)          TRUE    FALSE FALSE    0.70571   0   
#> 8  mybeh rate mybeh (period 2)          TRUE    FALSE FALSE    1.50000   0   
#> 9  mybeh mybeh linear shape             TRUE    FALSE FALSE    0.32237   0   
#> 10 mybeh mybeh quadratic shape          TRUE    FALSE FALSE    0.00000   0   
#>    effect1 effect2
#> 1   0        0    
#> 2   0        0    
#> 3   0        0    
#> 4   0        0    
#> 5   0        0    
#> 6  12      218    
#> 7   0        0    
#> 8   0        0    
#> 9   0        0    
#> 10  0        0