Function to perform a meta analysis of a collection of Siena fits.

Estimates a meta analysis based on a collection of Siena fits.

siena08(..., projname = "sienaMeta", bound = 5, alpha = 0.05, maxit=20)

Arguments

...
names of sienaFit objects, returned from siena07. They will be renamed if entered in format newname=oldname. It is also allowed to give for ... a list of sienaFit objects.

projname
Base name of report file if required

bound
Upper limit of standard error for inclusion in the meta analysis.

alpha
1 minus confidence level of confidence intervals.

maxit
Number of iterations of iterated least squares procedure.

Details

A meta analysis is performed as described in the Siena manual, section "Meta-analysis of Siena results". This consists of three parts: an iterated weighted least squares modification of the method described in the reference below; maximum likelihood estimates and confidence intervals based on profile likelihoods under normality assumptions; and Fisher combinations of left-sided and right-sided \(p\)-values. These are produced for all effects separately.

Note that the corresponding effects must have the same effect name in each model fit. This implies that at least covariates and behavior variables must have the same name in each model fit.

Value

An object of class sienaMeta. There are print, summary and plot methods for this class. This object contains at least the following.

thetadf

Data frame containing the coefficients, standard errors and score test results

projname

Name for any output file to be produced by the print method

bound

Estimates with standard error above this value were excluded from the calculations

scores

Object of class by indicating, for each effect in the models, whether score test information was present.

requestedEffects

The requestedEffects component of the first sienaFit object in ....

theta

The vector of ML estimates mu.ml (see below).

se

The vector of standard errors of the ML estimates mu.ml.se (see below).

Then for each effect, there is a list with at least the following.
cor.est

Spearman rank correlation coefficient between estimates and their standard errors.

cor.pval

p-value for above

regfit

Part of the result of the fit of iwlsm.

regsummary

The summary of the fit, which includes the coefficient table.

Tsq

test statistic for effect zero in every model

pTsq

p-value for above

tratio

test statistics that mean effect is 0

ptratio

p-value for above

Qstat

Test statistic for variance of effects is zero

pttilde

p-value for above

cjplus

Test statistic for at least one theta strictly greater than 0

cjminus

Test statistic for at least one theta strictly less than 0

cjplusp

p-value for cjplus

cjminusp

p-value for cjminus

mu.ml

ML estimate of population mean

mu.ml.se

standard error of ML estimate of population mean

sigma.ml

ML estimate of population standard deviation

mu.confint

confidence interval for population mean based on profile likelihood

sigma.confint

confidence interval for population standard deviation based on profile likelihood

n1

Number of fits on which the meta analysis is based

cjplus

Test statistic for combination of right one-sided Fisher combination tests

cjminus

Test statistic for combination of left one-sided Fisher combination tests

cjplusp

p-value for cjplus

cjminusp

p-value for cjminus

scoreplus

Test statistic for combination of right one-sided \(p\)-values from score tests

scoreminus

Test statistic for combination of left one-sided \(p\)-values from score tests

scoreplusp

p-value for scoreplus

scoreminusp

p-value for scoreminus

ns

Number of fits on which the score test analysis is based

References

T. A. B. Snijders and Chris Baerveldt (2003). Multilevel network study of the effects of delinquent behavior on friendship evolution. Journal of Mathematical Sociology 27, 123--151.

See also the manual (Section 11.2) and http://www.stats.ox.ac.uk/~snijders/siena/

See also

Examples

if (FALSE) {
# A meta-analysis for three groups does not make much sense
# for generalizing to a population of networks,
# but the Fisher combinations of p-values are meaningful.
# However, using three groups does show the idea.

Group1 <- sienaDependent(array(c(N3401, HN3401), dim=c(45, 45, 2)))
Group3 <- sienaDependent(array(c(N3403, HN3403), dim=c(37, 37, 2)))
Group4 <- sienaDependent(array(c(N3404, HN3404), dim=c(33, 33, 2)))
dataset.1 <- sienaDataCreate(Friends = Group1)
dataset.3 <- sienaDataCreate(Friends = Group3)
dataset.4 <- sienaDataCreate(Friends = Group4)
OneAlgorithm <- sienaAlgorithmCreate(projname = "SingleGroups", seed=128)
effects.1 <- getEffects(dataset.1)
effects.3 <- getEffects(dataset.3)
effects.4 <- getEffects(dataset.4)
effects.1 <- includeEffects(effects.1, transTrip)
effects.1 <- setEffect(effects.1, transRecTrip, fix=TRUE, test=TRUE)
effects.3 <- includeEffects(effects.3, transTrip)
effects.3 <- setEffect(effects.3, transRecTrip, fix=TRUE, test=TRUE)
effects.4 <- includeEffects(effects.4, transTrip)
effects.4 <- setEffect(effects.4, transRecTrip, fix=TRUE, test=TRUE)
ans.1 <- siena07(OneAlgorithm, data=dataset.1, effects=effects.1, batch=TRUE)
ans.3 <- siena07(OneAlgorithm, data=dataset.3, effects=effects.3, batch=TRUE)
ans.4 <- siena07(OneAlgorithm, data=dataset.4, effects=effects.4, batch=TRUE)
ans.1
ans.3
ans.4
(meta <- siena08(ans.1, ans.3, ans.4))
plot(meta, which=2:3, layout = c(2,1))
# For specifically presenting the Fisher combinations:
# First determine the components of meta with estimated effects:
which.est <- sapply(meta, function(x){ifelse(is.list(x),!is.null(x$cjplus),FALSE)})
Fishers <- t(sapply(1:sum(which.est),
        function(i){c(meta[[i]]$cjplus, meta[[i]]$cjminus,
                        meta[[i]]$cjplusp, meta[[i]]$cjminusp, 2*meta[[i]]$n1 )}))
Fishers <- as.data.frame(Fishers, row.names=names(meta)[which.est])
names(Fishers) <- c('Fplus', 'Fminus', 'pplus', 'pminus', 'df')
Fishers
round(Fishers,4)
}