Function to create a Siena data object

Creates a Siena data object from input dependent variables (networks and possibly behavioural variables), covariates, and composition change objects.

sienaDataCreate(..., nodeSets=NULL, getDocumentation=FALSE)

Arguments

...
objects of class sienaDependent, coCovar,varCovar,coDyadCovar,
varDyadCovar, sienaCompositionChange

nodeSets
list of Siena node sets. Default is the single node set named "Actors", length equal to the number of rows in the first object of class "sienaDependent". If the entire data set contains more than one node set, then the node sets must have been specified in the creation of all data objects.

getDocumentation
Flag to allow documentation of internal functions, not for use by users.

Details

The function checks that the objects fit, that there is at least one network, and adds various attributes to each dependent variable describing the data. If there is more than one nodeSet they must all be specified. Function print01Report will give a basic description of the data object and is a check useful, e.g., for diagnosing problems.

Value

An object of class "siena" which is designed to be used in a siena model fit. The components of the object are.

nodeSets

List of node sets involved

observations

Integer indicating number of waves of data

depvars

List of networks and behavior variables

cCovars

List of constant covariates

vCovars

List of changing covariates

dycCovars

List of constant dyadic covariates

dyvCovars

List of changing dyadic covariates

compositionChange

List of composition change objects corresponding to the node sets

References

See http://www.stats.ox.ac.uk/~snijders/siena/

See also

Examples

mynet <- sienaDependent(array(c(s501, s502, s503), dim=c(50, 50, 3)))
mybeh <- sienaDependent(s50a, type="behavior")
mydata <- sienaDataCreate(mynet, mybeh)
## And for a two-mode network
mynet1 <- sienaDependent(array(c(s501, s502), dim=c(50, 50, 2)), nodeSet="senders")
senders <- sienaNodeSet(50, nodeSetName="senders")
receivers <- sienaNodeSet(30, nodeSetName="receivers")
mynet2 <- sienaDependent(array(c(s501[,1:30], s502[,1:30]), dim=c(50, 30, 2)),
      nodeSet=c("senders", "receivers"))
(mydata <- sienaDataCreate(mynet1, mynet2, nodeSets=list(senders, receivers)))

#> Dependent variables:  mynet1, mynet2 
#> Number of observations: 2 
#> 
#> Nodesets                senders       receivers 
#> Number of nodes              50              30 
#> 
#> Dependent variable mynet1     
#> Type               oneMode    
#> Observations       2          
#> Nodeset            senders    
#> Densities          0.046 0.047
#> 
#> Dependent variable mynet2     
#> Type               bipartite  
#> Observations       2          
#> First nodeset      senders    
#> Second nodeset     receivers  
#> Densities          0.045 0.043
#> 

if (FALSE) {
print01Report(mydata, modelname = "mydescription")
}