Functions to assess the relative importance of effects at observation moments

The function sienaRI returns the relative importance of effects of a SAOM according to the measure of relative importance described in Section 3.1 of Indlekofer and Brandes (2013). The measure is based on the influence of effects on potential tie change or behavior change decisions of individual actors at the given observation moments. It takes the data as well as the complete model specification into account. Therefore, necessary arguments are the analysed data given as a siena data object as well as the complete model specification represented either by an estimated sienaFit object or by the triple consisting of a suitable parameter vector theta and the corresponding sienaAlgorithm and sienaEffects objects.
Entropy-based effect sizes as in Snijders (2004) and the within-ego standard deviations of change statistics are also computed. If getChangeStats=TRUE, the arrays of change statistics are stored in the sienaRI object.

sienaRI(data, ans=NULL, theta=NULL, algorithm=NULL, effects=NULL,
        getChangeStats=FALSE)
# S3 method for sienaRI
print(x, printSigma=FALSE,...)
# S3 method for sienaRI
plot(x, actors = NULL, col = NULL, addPieChart = FALSE,
              radius = 1, width = NULL, height = NULL, legend = TRUE,
              legendColumns = NULL, legendHeight = NULL,
              cex.legend = NULL, cex.names = NULL,...)

Arguments

data
siena data object representing the analyzed data and resulting from a call to sienaDataCreate.

ans
sienaFit object resulting from a call to siena07. The sienaFit object contains all necessary information on the model specification, in particular, the vector of parameter values ans$theta, the used algorithm for estimation ans$x, and information on included model effects ans$effects. If ans is a valid sienaFit object, the calculations of relative importances are based on ans$theta, ans$x, and ans$effects. Alternatively, the necessary information can be given directly as a suitable parameter vector theta, a sienaAlgorithm object, and a sienaEffects object. In this case, ans has to be unspecified (i.e., ans=NULL).

theta
Vector of parameter values of effects included in the model. Length of theta has to be equal to the number of included effects.

algorithm
sienaAlgorithm object as resulting from a call to sienaAlgorithmCreate. Works only for estimation by Method of Moments (i.e., if maxlike = FALSE).

effects
sienaEffects object specifying which effects are included the model. Note that sienaRI does not yet work for endowment or creation effect, i.e., included effects have to be of type eval (or rate).

getChangeStats
Boolean: If TRUE, an array of change statistics is added to the sienaRI object.

x
sienaRI object resulting from a call to sienaRI.

printSigma
Boolean: If TRUE, average within-ego standard deviations of change statistics ('sigma'), are included in the print.

actors
vector of integers: set of actors to be included in the plot; if NULL, all actors.

col
Colors used in the plot. If col=NULL a default color scheme is used.

addPieChart
Boolean: If TRUE, pie charts of aggregated relative importances for the complete set of actors will be added to the plot.

radius
Radius of pie charts. Only effective if addPieCharts = TRUE.

width
Width of the plot. If width=NULL a default value is used.

height
Height of the plot. If height=NULL a default value is used.

legend
Boolean: if TRUE a legend is added to the plot. x$effectNames are used as labels.

legendColumns
Number of columns in legend. If legendColumns=NULL a default value is used. Only effective if legend=TRUE.

legendHeight
Height of legend. If legendHeight=NULL a default value is used. Only effective if legend=TRUE.

cex.legend
Specifies the relative font size of legend labels.

cex.names
Specifies the relative font size of bar graph labels.

...
Other arguments.

Details

sienaRI takes the data as well as the complete model specification into account. Therefore, necessary arguments are the analyzed data given as a siena data object as well as the complete model specification represented either by an estimated sienaFit object or by the triple consisting of a suitable parameter vector theta and the corresponding sienaAlgorithm and sienaEffects objects.

A sienaFit object contains all necessary information on the model specification, in particular, the vector of parameter values ans$theta, the used algorithm for estimation ans$x, and information on included model effects ans$effects.

If ans is a valid sienaFit object, the calculations of relative importances are based on ans$theta, ans$x, and ans$effects. Alternatively, the necessary information can be given directly as a suitable parameter vector theta, a sienaAlgorithm object, and a sienaEffects object. In this case, ans has to be unspecified, i.e., ans=NULL.

Note that sienaRI works only with Method of Moments (i.e., for sienaAlgorithm objects with maxlike = FALSE) and that it does not yet work for endowment or creation effects (i.e., included effects have to be of type eval), and also not for models with interaction effects. It does not allow two-mode (bipartite) networks as dependent variables; but these can be represented as one-mode networks using structural zeros. If the network is non-directed, the relative importances and entropy-based 'degrees of certainty' are calculated for modelType=2 ('forcing'; see sienaAlgorithmCreate).

If there are any missing tie values in the network data set, they are imputed by initial zeros and Last Observation Carried Forward. Structural zeros and ones are replaced by NA and treated as impossible choices in the probability vectors and ignored in the standard deviations; but the change statistics for these dyads still are given in changeStatistics (if requested).

Value

If the model contains only one dependent variable, sienaRI returns an object of class sienaRI. Otherwise, it returns a list of objects of class sienaRI, each corresponding to one dependent variable.

A returned sienaRI object stores the expected relative importances of effects of one dependent variable at observation moments as defined in Section 3.1 of Indlekofer and Brandes (2013).

A sienaRI object is a list with the following components. For the components referred to as lists themselves, these are lists corresponding to the observation moments.

- dependentVariable

the name of the corresponding dependent variable.

- effectNames

the names of considered effects.

- RIActors

a list that contains the expected relative importances of effects for each potential actor decision at observation moments. This is equation (3) in Indlekofer and Brandes (2013).

- expectedRI

a list that contains the expected relative importances of effects aggregated over all actors for each network observation. These are the averages of the actor related values in RIActors. This is equation (4) in Indlekofer and Brandes (2013).

- IActors

a list that contains the expected importances of effects for each potential actor decision at observation moments. This is the numerator of equation (3) in Indlekofer and Brandes (2013).

- expectedI

is a list that contains the expected importances of effects aggregated over all actors in each observation. More precisely, it contains the averages of the actor related values in IActors.

- absoluteSumActors

a list that contains the sum of the (unstandardized) L1-differences calculated for each potential actor decision at observation moments. This is the denominator of equation (3) in Indlekofer and Brandes (2013).

- RHActors

a list that contains the degree of certainty in the potential ministep taken by an actor at the observation moments; this is R_H(i,x) of formula (6) in Snijders (2004). The mean over actors of these degrees of certainty, given by formula (7) in Snijders (2004), is printed by the print method for sienaRI objects.

- sigma

a list of effects by ego matrices of the values of the within-ego standard deviations of the change statistics. Their averages (over egos) are printed if printSigma=TRUE.

- changeStatistics

a list of arrays (effects by alters by egos) containing for each observation moment, the values of the change statistics for toggling the tie from actor to ego; this is produced only if getChangeStats=TRUE.

References

Indlekofer, Natalie and Brandes, Ulrik (2013). Relative Importance of Effects in Stochastic Actor-oriented Models. Network Science, 1 (3), 278-304.

Snijders, Tom A.B. (2004). Explained Variation in Dynamic Network Models. Mathematics and Social Sciences, 168 (4), 31-41.

Examples

myalgorithm <- sienaAlgorithmCreate(nsub=1, n3=50, projname=NULL)

#> If you use this algorithm object, siena07 will create an output file /var/folders/pl/ngry9kn13p7466xwk0_2xyh40000gn/T//RtmpBuM5HN/Siena344e123fac5.txt .
#> This is a temporary file for this R session.

mynet1 <- sienaDependent(array(c(tmp3, tmp4), dim=c(32, 32, 2)))
mydata <- sienaDataCreate(mynet1)
myeff <- getEffects(mydata)
myeff <- includeEffects(myeff, density, recip, transTies, nbrDist2)

#>   effectName                 include fix   test  initialValue parm
#> 1 outdegree (density)        TRUE    FALSE FALSE   -0.56039   0   
#> 2 reciprocity                TRUE    FALSE FALSE    0.00000   0   
#> 3 transitive ties            TRUE    FALSE FALSE    0.00000   0   
#> 4 number of actors at dist 2 TRUE    FALSE FALSE    0.00000   0   

ans <- siena07(myalgorithm, data=mydata, effects=myeff, batch=TRUE)

#> 
#> Start phase 0 
#> theta: -0.56  0.00  0.00  0.00 
#> 
#> Start phase 1 
#> Phase 1 Iteration 1 Progress: 0%
#> Phase 1 Iteration 2 Progress: 0%
#> Phase 1 Iteration 3 Progress: 1%
#> Phase 1 Iteration 4 Progress: 1%
#> Phase 1 Iteration 5 Progress: 1%
#> Phase 1 Iteration 10 Progress: 2%
#> Phase 1 Iteration 15 Progress: 3%
#> Phase 1 Iteration 20 Progress: 4%
#> Phase 1 Iteration 25 Progress: 5%
#> Phase 1 Iteration 30 Progress: 6%
#> Phase 1 Iteration 35 Progress: 7%
#> Phase 1 Iteration 40 Progress: 8%
#> Phase 1 Iteration 45 Progress: 9%
#> Phase 1 Iteration 50 Progress: 9%
#> theta: -0.7220  0.1963  0.1335 -0.0184 
#> 
#> Start phase 2.1
#> Phase 2 Subphase 1 Iteration 1 Progress: 48%
#> Phase 2 Subphase 1 Iteration 2 Progress: 48%
#> theta -0.850  0.576  0.137 -0.111 
#> ac -6.447  1.011  0.115  2.788 
#> Phase 2 Subphase 1 Iteration 3 Progress: 48%
#> Phase 2 Subphase 1 Iteration 4 Progress: 48%
#> theta -1.213  1.473  0.489 -0.258 
#> ac -1.509  1.261  0.453  2.545 
#> Phase 2 Subphase 1 Iteration 5 Progress: 48%
#> Phase 2 Subphase 1 Iteration 6 Progress: 49%
#> theta -1.584  2.160  0.809 -0.321 
#> ac -0.848  1.279  0.476  1.713 
#> Phase 2 Subphase 1 Iteration 7 Progress: 49%
#> Phase 2 Subphase 1 Iteration 8 Progress: 49%
#> theta -1.76  2.43  1.06 -0.33 
#> ac -0.710  1.204  0.475  1.710 
#> Phase 2 Subphase 1 Iteration 9 Progress: 49%
#> Phase 2 Subphase 1 Iteration 10 Progress: 49%
#> theta -1.968  2.722  1.253 -0.352 
#> ac -0.689  1.205  0.429  1.509 
#> theta -2.312  2.109  1.587 -0.263 
#> ac -0.10903 -0.01085 -0.00189 -0.21865 
#> theta: -2.312  2.109  1.587 -0.263 
#> 
#> Start phase 3 

RI <- sienaRI(mydata, ans)
RI

#> 
#>   Expected relative importance of effects for dependent variable 'mynet1' at observation moments:
#> 
#> 
#>                                                                wave 1    wave 2    
#>  
#>   1. eval outdegree (density)                                  0.3608    0.3204  
#>   2. eval reciprocity                                          0.2092    0.2082  
#>   3. eval transitive ties                                      0.1838    0.2328  
#>   4. eval number of actors at dist 2                           0.2461    0.2386  
#> 
#> 
#>   Expected importance of effects for this dependent variable:
#> 
#>   1. eval outdegree (density)                                  0.8912    0.7588  
#>   2. eval reciprocity                                          0.5205    0.5083  
#>   3. eval transitive ties                                      0.5439    0.6512  
#>   4. eval number of actors at dist 2                           0.5837    0.5012  
#> 
#> 
#>  R_H ('degree of certainty')                                   0.2875    0.2600  

if (FALSE) {
plot(RI, addPieChart=TRUE)
plot(RI, actors=1:20, addPieChart=TRUE, radius=1.08)
}

myalgorithm <- sienaAlgorithmCreate(nsub=1, n3=50, projname=NULL)

#> If you use this algorithm object, siena07 will create an output file /var/folders/pl/ngry9kn13p7466xwk0_2xyh40000gn/T//RtmpBuM5HN/Siena344e61249c18.txt .
#> This is a temporary file for this R session.

mynet2 <- sienaDependent(array(c(s502, s503), dim=c(50, 50, 2)))
mybeh <- sienaDependent(s50a[,2:3], type="behavior")
mydata2 <- sienaDataCreate(mynet2, mybeh)
myeff2 <- getEffects(mydata2)
myeff2 <- includeEffects(myeff2, density, recip, transTies)

#>   effectName          include fix   test  initialValue parm
#> 1 outdegree (density) TRUE    FALSE FALSE   -1.44421   0   
#> 2 reciprocity         TRUE    FALSE FALSE    0.00000   0   
#> 3 transitive ties     TRUE    FALSE FALSE    0.00000   0   

ans2 <- siena07(myalgorithm, data=mydata2, effects=myeff2, batch=TRUE)

#> 
#> Start phase 0 
#> theta:  4.329 -1.444  0.000  0.000  0.849  0.316  0.000 
#> 
#> Start phase 1 
#> Phase 1 Iteration 1 Progress: 0%
#> Phase 1 Iteration 2 Progress: 0%
#> Phase 1 Iteration 3 Progress: 0%
#> Phase 1 Iteration 4 Progress: 1%
#> Phase 1 Iteration 5 Progress: 1%
#> Phase 1 Iteration 10 Progress: 1%
#> Phase 1 Iteration 15 Progress: 2%
#> Phase 1 Iteration 20 Progress: 3%
#> Phase 1 Iteration 25 Progress: 4%
#> Phase 1 Iteration 30 Progress: 4%
#> Phase 1 Iteration 35 Progress: 5%
#> Phase 1 Iteration 40 Progress: 6%
#> Phase 1 Iteration 45 Progress: 7%
#> Phase 1 Iteration 50 Progress: 7%
#> theta:  4.6577 -1.6610  0.6993  0.3784  0.9478  0.0179 -0.1245 
#> 
#> Start phase 2.1
#> Phase 2 Subphase 1 Iteration 1 Progress: 59%
#> Phase 2 Subphase 1 Iteration 2 Progress: 59%
#> theta  4.5403 -1.8023  1.1788  0.5811  1.0446  0.0107 -0.1978 
#> ac 42.005  0.728  2.088  1.983  2.408  0.995  1.137 
#> Phase 2 Subphase 1 Iteration 3 Progress: 59%
#> Phase 2 Subphase 1 Iteration 4 Progress: 59%
#> theta  4.525 -2.219  2.157  0.959  1.384  0.177 -0.357 
#> ac  1.3792  1.0977  1.9790  2.1497 -0.0941  1.0925  1.0673 
#> Phase 2 Subphase 1 Iteration 5 Progress: 59%
#> Phase 2 Subphase 1 Iteration 6 Progress: 59%
#> theta  4.203 -2.561  2.444  1.112  1.533  0.622 -0.449 
#> ac  0.3548  1.3076  2.0261  2.3603 -0.0998  0.9877  1.0001 
#> Phase 2 Subphase 1 Iteration 7 Progress: 59%
#> Phase 2 Subphase 1 Iteration 8 Progress: 60%
#> theta  4.340 -2.773  2.624  1.200  1.583  0.666 -0.522 
#> ac  0.0888  1.3582  1.8451  2.4110 -0.1430  0.9876  1.0000 
#> Phase 2 Subphase 1 Iteration 9 Progress: 60%
#> Phase 2 Subphase 1 Iteration 10 Progress: 60%
#> theta  4.697 -2.932  2.723  1.185  1.605  0.771 -0.554 
#> ac 0.242 1.380 1.977 2.564 0.399 0.526 0.701 
#> theta  6.049 -3.160  2.808  1.375  1.470  0.449 -0.449 
#> ac -0.1649 -0.0238 -0.4546 -0.4560 -0.0737 -0.1076 -0.1140 
#> theta:  6.049 -3.160  2.808  1.375  1.470  0.449 -0.449 
#> 
#> Start phase 3 

# Use only the parameters for the evaluation function:
theta.eval <- ans2$theta[myeff2$type[myeff2$include]=="eval"]
RI <- sienaRI(mydata2, theta=theta.eval, algorithm=myalgorithm,
              effects = myeff2)

#> more than one dependent variable
#>  return value is therefore not of class 'sienaRI'
#>  but a list of objects of class 'sienaRI'.

RI

#> [[1]]
#> 
#>   Expected relative importance of effects for dependent variable 'mynet2' at observation moments:
#> 
#> 
#>                                                                wave 1    wave 2    
#>  
#>   1. eval outdegree (density)                                  0.4953    0.4652  
#>   2. eval reciprocity                                          0.2939    0.3390  
#>   3. eval transitive ties                                      0.2107    0.1957  
#> 
#> 
#>   Expected importance of effects for this dependent variable:
#> 
#>   1. eval outdegree (density)                                  1.0942    1.0232  
#>   2. eval reciprocity                                          0.7579    0.8664  
#>   3. eval transitive ties                                      0.5766    0.5316  
#> 
#> 
#>  R_H ('degree of certainty')                                   0.4699    0.4263  
#> 
#> [[2]]
#> 
#>   Expected relative importance of effects for dependent variable 'mybeh' at observation moments:
#> 
#> 
#>                                                                wave 1    wave 2    
#>  
#>   1. eval mybeh linear shape                                   0.3822    0.3782  
#>   2. eval mybeh quadratic shape                                0.6178    0.6218  
#> 
#> 
#>   Expected importance of effects for this dependent variable:
#> 
#>   1. eval mybeh linear shape                                   0.1955    0.1906  
#>   2. eval mybeh quadratic shape                                0.4088    0.4106  
#> 
#> 
#>  R_H ('degree of certainty')                                   0.1479    0.1531  
#> 

if (FALSE) {
plot(RI[[2]], col = c("red", "green"), legend=FALSE)
plot(RI[[1]], addPieChart = TRUE, legendColumns=2)
}