Functions to assess and account for time heterogeneity of parameters

Takes a sienaFit object estimated by Method of Moments, and tests for time heterogeneity by the addition of interactions with time dummy variables at waves m=2...(M-1). The test used is the score-type test of Schweinberger (2012). Tests for joint significance, parameter-wise significance, period-wise significance, individual significance, and one-step estimates of the unrestricted model parameters are returned in a list.

sienaTimeTest(sienaFit, effects=NULL, excludedEffects=NULL, condition=FALSE)

Arguments

sienaFit
A sienaFit object returned by siena07.

effects
Optional vector of effect numbers to test. Use the numbering on the print of the sienaFit object.

excludedEffects
Optional vector of effect numbers for which time heterogeneity is not to be tested. Use the numbering on the print of the sienaFit object.

condition
Whether to orthogonalize effect-wise score-type tests and individual significance tests against estimated effects and un-estimated dummy terms, or just against estimated effects.

Details

This test follows the score type test of Schweinberger (2012) as elaborated by Lospinoso et al. (2011) by using statistics already calculated at each wave to obtain vectors of partitioned moment functions corresponding to a restricted model (the model in the sienaFit object; used as null hypothesis) and an unrestricted model (which contains dummies for waves m=2...(M-1); used as alternative hypothesis).

condition=TRUE leads to a rough-and-easy approximation to controlling the mentioned tests also for the unestimated effects.

After assessing time heterogeneity, effects objects can be modified by adding numbers of all or some periods to the timeDummy column. This is facilitated by the includeTimeDummy function. For an effects object in which the timeDummy column of some of the included effects includes some or all period numbers, interactions of those effects with time dummies for the indicated periods will also be estimated.

An alternative to the use of includeTimeDummy is to define time-dependent actor covariates (dummy variables or other functions of wave number that are the same for all actors), include these in the data set through sienaAlgorithmCreate, and include interactions of other effects with ego effects of these time-dependent actor covariates by includeInteraction. This is illustrated in an example below. Using includeTimeDummy is easier; using self-defined interactions with time-dependent variables gives more control.

If you wish to use this function with sienaFit objects that use the finite differences method of derivative estimation, or which use maximum likelihood estimation, you must request the derivatives to be returned by wave using the byWave=TRUE option for siena07.

Effects leading to dummy interactions that are collinear with the model originally fitted, after excluding the effects mentioned, will be automatically excluded from the time heterogeneity testing.

If sienaTimeTest gives errors that there are too many collinear effects, run it with a smaller set of effects as specified by the effects parameter. For example, if the model has 40 effects of which the first 8 are rate parameters and therefore uninteresting, and there is such an error message, try effects=9:30; if that still does not work, decrease the upper limit of 30, if it does work increase it, to find the largest possible set of effects for which heterogeneity assessment still is possible; then as a next step try the remaining effects in a similar way.

Also if the execution is time-consuming, e.g., for a multi-group sienaFit object with many groups and many effects, it can be helpful to carry out the function in smaller subsets of effects.

Value

sienaTimeTest returns a list containing many items, including the following:

JointTest

A chi-squared test for joint significance of the dummies.

EffectTest

A chi-squared test for joint significance across dummies for each separate effect.

GroupTest

A chi-squared test for joint significance across dummies; if sienaFit is a fit for a multi-group object then these refer to each group; else they refer to ecah period.

IndividualTest

A matrix displaying initial estimates, one-step estimates, and \(p\)-values for the individual interactions.

References

See http://www.stats.ox.ac.uk/~snijders/siena/ for general information on RSiena.

J.A. Lospinoso, M. Schweinberger, T.A.B. Snijders, and R.M. Ripley (2011). Assessing and Accounting for Time Heterogeneity in Stochastic Actor Oriented Models. Advances in Data Analysis and Computation, 5:147-176.

M. Schweinberger (2012). Statistical modeling of network panel data: Goodness-of-fit. British Journal of Statistical and Mathematical Psychology 65, 263--281.

See also

Examples

## Estimate a restricted model
myalgorithm <- sienaAlgorithmCreate(nsub=1, n3=50, projname=NULL)

#> If you use this algorithm object, siena07 will create an output file /var/folders/pl/ngry9kn13p7466xwk0_2xyh40000gn/T//RtmpBuM5HN/Siena344e2a841d7b.txt .
#> This is a temporary file for this R session.

# Short estimation not for practice, just for having a quick demonstration
mynet1 <- sienaDependent(array(c(s501, s502, s503), dim=c(50, 50, 3)))
mydata <- sienaDataCreate(mynet1)
myeff <- getEffects(mydata)
myeff <- includeEffects(myeff, transTrip)

#>   effectName          include fix   test  initialValue parm
#> 1 transitive triplets TRUE    FALSE FALSE          0   0   

ans <- siena07(myalgorithm, data=mydata, effects=myeff, batch=TRUE)

#> 
#> Start phase 0 
#> theta: -1.47  0.00  0.00 
#> 
#> Start phase 1 
#> Phase 1 Iteration 1 Progress: 0%
#> Phase 1 Iteration 2 Progress: 0%
#> Phase 1 Iteration 3 Progress: 1%
#> Phase 1 Iteration 4 Progress: 1%
#> Phase 1 Iteration 5 Progress: 1%
#> Phase 1 Iteration 10 Progress: 2%
#> Phase 1 Iteration 15 Progress: 3%
#> Phase 1 Iteration 20 Progress: 4%
#> Phase 1 Iteration 25 Progress: 5%
#> Phase 1 Iteration 30 Progress: 6%
#> Phase 1 Iteration 35 Progress: 7%
#> Phase 1 Iteration 40 Progress: 8%
#> Phase 1 Iteration 45 Progress: 9%
#> Phase 1 Iteration 50 Progress: 11%
#> theta: -1.732  0.319  0.257 
#> 
#> Start phase 2.1
#> Phase 2 Subphase 1 Iteration 1 Progress: 42%
#> Phase 2 Subphase 1 Iteration 2 Progress: 43%
#> theta -1.821  0.480  0.363 
#> ac 1.64 1.12 1.08 
#> Phase 2 Subphase 1 Iteration 3 Progress: 43%
#> Phase 2 Subphase 1 Iteration 4 Progress: 43%
#> theta -2.111  0.984  0.700 
#> ac 0.357 1.238 0.817 
#> Phase 2 Subphase 1 Iteration 5 Progress: 43%
#> Phase 2 Subphase 1 Iteration 6 Progress: 43%
#> theta -2.394  1.716  0.743 
#> ac 0.541 1.233 0.745 
#> Phase 2 Subphase 1 Iteration 7 Progress: 44%
#> Phase 2 Subphase 1 Iteration 8 Progress: 44%
#> theta -2.537  2.206  0.689 
#> ac 0.252 1.159 0.522 
#> Phase 2 Subphase 1 Iteration 9 Progress: 44%
#> Phase 2 Subphase 1 Iteration 10 Progress: 44%
#> theta -2.712  2.531  0.685 
#> ac 0.261 1.100 0.377 
#> theta -2.696  2.464  0.621 
#> ac -0.24270 -0.00443 -0.41906 
#> theta: -2.696  2.464  0.621 
#> 
#> Start phase 3 

## Conduct the score-type test to assess whether heterogeneity is present.
tt <- sienaTimeTest(ans)
summary(tt)

#> Joint significance test of time heterogeneity:
#> chi-squared =   3.69, d.f. = 3, p= 0.2973, 
#> where H0: The following parameters are zero:
#>  (1) (*)Dummy2:outdegree (density)
#>  (2) (*)Dummy2:reciprocity
#>  (3) (*)Dummy2:transitive triplets
#> 
#> Individual significance tests and one-step estimators:
#>                               Initial Est. One Step Est. p-Value
#> outdegree (density)           -2.6960      -2.5004        0.0000
#> reciprocity                    2.4642       2.2196        0.0000
#> transitive triplets            0.6209       0.5629        0.0000
#> (*)Dummy2:outdegree (density)  0.0000      -0.3707        0.6020
#> (*)Dummy2:reciprocity          0.0000       0.4797        0.1910
#> (*)Dummy2:transitive triplets  0.0000       0.1453        0.3200
#> 
#> Effect-wise joint significance tests
#> (i.e. each effect across all dummies):
#>                     chi-sq. df p-value
#> outdegree (density)    0.27  1   0.603
#> reciprocity            1.71  1   0.191
#> transitive triplets    0.99  1   0.320
#> 
#> Period-wise joint significance tests
#> (i.e. each period across all parameters):
#>          chi-sq. df p-value
#> Period 1    3.69  3   0.297
#> Period 2    3.69  3   0.297
#> 
#> Use the following indices for plotting:
#>   (1) outdegree (density)
#>   (2) reciprocity
#>   (3) transitive triplets
#> 
#> If you would like to fit time dummies to your model,
#> use the includeTimeDummy function.
#> Type "?sienaTimeTest" for more information on this output.

## Suppose that we wish to include time dummies.
## Add them in the following way:
myeff <- includeTimeDummy(myeff, recip, transTrip, timeDummy="2")

#>                     name          effectName                  functionName
#> mynet1.obj.eval.2 mynet1         reciprocity   Number of reciprocated ties
#> mynet1.obj.eval.3 mynet1 transitive triplets Number of transitive triplets
#>                   shortName interaction1 interaction2 type basicRate include
#> mynet1.obj.eval.2     recip                           eval     FALSE    TRUE
#> mynet1.obj.eval.3 transTrip                           eval     FALSE    TRUE
#>                   randomEffects   fix  test timeDummy initialValue parm
#> mynet1.obj.eval.2         FALSE FALSE FALSE         2            0    0
#> mynet1.obj.eval.3         FALSE FALSE FALSE         2            0    0
#>                   functionType period rateType untrimmedValue effect1 effect2
#> mynet1.obj.eval.2    objective   <NA>     <NA>              0       0       0
#> mynet1.obj.eval.3    objective   <NA>     <NA>              0       0       0
#>                   effect3 interactionType local setting effectFn statisticFn
#> mynet1.obj.eval.2       0          dyadic  TRUE             NULL        NULL
#> mynet1.obj.eval.3       0                  TRUE             NULL        NULL
#>                   netType groupName group effectNumber
#> mynet1.obj.eval.2 oneMode    Group1     1           11
#> mynet1.obj.eval.3 oneMode    Group1     1           14

ans2 <- siena07(myalgorithm, data=mydata, effects=myeff, batch=TRUE)

#> 
#> Start phase 0 
#> theta: -1.47  0.00  0.00  0.00  0.00  0.00 
#> 
#> Start phase 1 
#> Phase 1 Iteration 1 Progress: 0%
#> Phase 1 Iteration 2 Progress: 0%
#> Phase 1 Iteration 3 Progress: 0%
#> Phase 1 Iteration 4 Progress: 1%
#> Phase 1 Iteration 5 Progress: 1%
#> Phase 1 Iteration 10 Progress: 2%
#> Phase 1 Iteration 15 Progress: 2%
#> Phase 1 Iteration 20 Progress: 3%
#> Phase 1 Iteration 25 Progress: 4%
#> Phase 1 Iteration 30 Progress: 5%
#> Phase 1 Iteration 35 Progress: 6%
#> Phase 1 Iteration 40 Progress: 6%
#> Phase 1 Iteration 45 Progress: 7%
#> Phase 1 Iteration 50 Progress: 8%
#> theta: -1.7382  0.2963  0.3177  0.0000  0.0666  0.0132 
#> 
#> Start phase 2.1
#> Phase 2 Subphase 1 Iteration 1 Progress: 56%
#> Phase 2 Subphase 1 Iteration 2 Progress: 56%
#> theta -1.8534  0.4871  0.4668  0.0000  0.1150  0.0477 
#> ac 0.563 1.144 1.496 2.226 3.173 6.592 
#> Phase 2 Subphase 1 Iteration 3 Progress: 56%
#> Phase 2 Subphase 1 Iteration 4 Progress: 56%
#> theta -2.23499  1.22215  0.88051  0.00000  0.33326 -0.00267 
#> ac 0.494 1.237 0.344 3.803 0.913 1.251 
#> Phase 2 Subphase 1 Iteration 5 Progress: 56%
#> Phase 2 Subphase 1 Iteration 6 Progress: 56%
#> theta -2.427  1.992  0.736  0.000  0.606 -0.246 
#> ac  0.528  0.862  0.270 -2.449 -0.925 -1.047 
#> Phase 2 Subphase 1 Iteration 7 Progress: 56%
#> Phase 2 Subphase 1 Iteration 8 Progress: 57%
#> theta -2.502  2.381  0.523  0.000  0.454  0.113 
#> ac  0.4960  0.8197  0.2465  0.0921 -0.4942 -1.2642 
#> Phase 2 Subphase 1 Iteration 9 Progress: 57%
#> Phase 2 Subphase 1 Iteration 10 Progress: 57%
#> theta -2.5983  2.5943  0.5094  0.0000  0.3431 -0.0993 
#> ac  0.416  0.746  0.117  0.138 -0.525 -1.363 
#> theta -2.7214  2.5379  0.6212  0.0000  0.2160  0.0321 
#> ac -0.33060 -0.03012 -0.52706  0.00564 -0.42709 -0.39023 
#> theta: -2.7214  2.5379  0.6212  0.0000  0.2160  0.0321 
#> 
#> Start phase 3 

## Re-assess the time heterogeneity
(tt2 <- sienaTimeTest(ans2))

#> Joint significance test of time heterogeneity:
#> chi-squared =   2.11, d.f. = 1, p= 0.1461, 
#> where H0: The following parameters are zero:
#>  (1) (*)Dummy2:outdegree (density)

## And so on..

## A demonstration of the plotting facilities, on a larger dataset:
## (Of course pasting these identical sets of three waves after each other
##  in a sequence of six is not really meaningful. It's just a demonstration.)

myalgorithm <- sienaAlgorithmCreate(nsub=2, n3=50, seed=654, projname=NULL)

#> If you use this algorithm object, siena07 will create an output file /var/folders/pl/ngry9kn13p7466xwk0_2xyh40000gn/T//RtmpBuM5HN/Siena344e47ab8c0b.txt .
#> This is a temporary file for this R session.

mynet1 <- sienaDependent(array(c(s501, s502, s503, s501, s503, s502),
                               dim=c(50, 50, 6)))
mydata <- sienaDataCreate(mynet1)
myeff <- getEffects(mydata)
myeff <- includeEffects(myeff, transTrip, balance)

#>   effectName          include fix   test  initialValue parm
#> 1 transitive triplets TRUE    FALSE FALSE          0   0   
#> 2 balance             TRUE    FALSE FALSE          0   0   

myeff <- includeTimeDummy(myeff, density, timeDummy="all")

#>                     name          effectName   functionName shortName
#> mynet1.obj.eval.1 mynet1 outdegree (density) Number of ties   density
#>                   interaction1 interaction2 type basicRate include
#> mynet1.obj.eval.1                           eval     FALSE    TRUE
#>                   randomEffects   fix  test timeDummy initialValue parm
#> mynet1.obj.eval.1          TRUE FALSE FALSE       all    -1.490548    0
#>                   functionType period rateType untrimmedValue effect1 effect2
#> mynet1.obj.eval.1    objective   <NA>     <NA>      -1.490548       0       0
#>                   effect3 interactionType local setting effectFn statisticFn
#> mynet1.obj.eval.1       0          dyadic  TRUE             NULL        NULL
#>                   netType groupName group effectNumber
#> mynet1.obj.eval.1 oneMode    Group1     1           11

myeff <- includeTimeDummy(myeff, recip, timeDummy="2,3,5")

#>                     name  effectName                functionName shortName
#> mynet1.obj.eval.2 mynet1 reciprocity Number of reciprocated ties     recip
#>                   interaction1 interaction2 type basicRate include
#> mynet1.obj.eval.2                           eval     FALSE    TRUE
#>                   randomEffects   fix  test timeDummy initialValue parm
#> mynet1.obj.eval.2         FALSE FALSE FALSE     2,3,5            0    0
#>                   functionType period rateType untrimmedValue effect1 effect2
#> mynet1.obj.eval.2    objective   <NA>     <NA>              0       0       0
#>                   effect3 interactionType local setting effectFn statisticFn
#> mynet1.obj.eval.2       0          dyadic  TRUE             NULL        NULL
#>                   netType groupName group effectNumber
#> mynet1.obj.eval.2 oneMode    Group1     1           14

myeff <- includeTimeDummy(myeff, balance, timeDummy="4")

#>                      name effectName      functionName shortName interaction1
#> mynet1.obj.eval.12 mynet1    balance Amount of balance   balance             
#>                    interaction2 type basicRate include randomEffects   fix
#> mynet1.obj.eval.12              eval     FALSE    TRUE         FALSE FALSE
#>                     test timeDummy initialValue parm functionType period
#> mynet1.obj.eval.12 FALSE         4            0    0    objective   <NA>
#>                    rateType untrimmedValue effect1 effect2 effect3
#> mynet1.obj.eval.12     <NA>              0       0       0       0
#>                    interactionType local setting effectFn statisticFn netType
#> mynet1.obj.eval.12                 FALSE             NULL        NULL oneMode
#>                    groupName group effectNumber
#> mynet1.obj.eval.12    Group1     1           42

if (FALSE) {
(ansp <- siena07(myalgorithm, data=mydata, effects=myeff, batch=TRUE))
ttp <- sienaTimeTest(ansp)

## Pairwise plots show
plot(ttp, pairwise=TRUE)

## Time test plots show
plot(ttp, effects=1:4, dims=c(2,2))
}

## Instead of working with includeTimeDummy,
## you can also define time dummies explicitly;
## this may give more control and more clarity:
dum2 <- matrix(c(0,1,0,0,0), nrow=50, ncol=5, byrow=TRUE)
dum3 <- matrix(c(0,0,1,0,0), nrow=50, ncol=5, byrow=TRUE)
dum4 <- matrix(c(0,0,0,1,0), nrow=50, ncol=5, byrow=TRUE)
dum5 <- matrix(c(0,0,0,0,1), nrow=50, ncol=5, byrow=TRUE)
time2 <- varCovar(dum2)
time3 <- varCovar(dum3)
time4 <- varCovar(dum4)
time5 <- varCovar(dum5)
mydata <- sienaDataCreate(mynet1, time2, time3, time4, time5)
myeff <- getEffects(mydata)
myeff <- includeEffects(myeff, transTrip, balance)

#>   effectName          include fix   test  initialValue parm
#> 1 transitive triplets TRUE    FALSE FALSE          0   0   
#> 2 balance             TRUE    FALSE FALSE          0   0   

## corresponding to includeTimeDummy(myeff, density, timeDummy="all"):
myeff <- includeEffects(myeff, egoX, interaction1='time2')

#>   effectName include fix   test  initialValue parm
#> 1 time2 ego  TRUE    FALSE FALSE          0   0   

myeff <- includeEffects(myeff, egoX, interaction1='time3')

#>   effectName include fix   test  initialValue parm
#> 1 time3 ego  TRUE    FALSE FALSE          0   0   

myeff <- includeEffects(myeff, egoX, interaction1='time4')

#>   effectName include fix   test  initialValue parm
#> 1 time4 ego  TRUE    FALSE FALSE          0   0   

myeff <- includeEffects(myeff, egoX, interaction1='time5')

#>   effectName include fix   test  initialValue parm
#> 1 time5 ego  TRUE    FALSE FALSE          0   0   

## corresponding to myeff <- includeTimeDummy(myeff, recip, timeDummy="2,3,5"):
myeff <- includeInteraction(myeff, egoX, recip, interaction1=c('time2', ''))

#>   effectName                     include fix   test  initialValue parm effect1
#> 1 unspecified interaction effect TRUE    FALSE FALSE          0   0    185    
#>   effect2
#> 1 18     

myeff <- includeInteraction(myeff, egoX, recip, interaction1=c('time3', ''))

#>   effectName                     include fix   test  initialValue parm effect1
#> 1 unspecified interaction effect TRUE    FALSE FALSE          0   0    339    
#>   effect2
#> 1 18     

myeff <- includeInteraction(myeff, egoX, recip, interaction1=c('time5', ''))

#>   effectName                     include fix   test  initialValue parm effect1
#> 1 unspecified interaction effect TRUE    FALSE FALSE          0   0    647    
#>   effect2
#> 1 18     

## corresponding to myeff <- includeTimeDummy(myeff, balance, timeDummy="4"):
myeff <- includeInteraction(myeff, egoX, balance, interaction1=c('time4', ''))

#>   effectName                     include fix   test  initialValue parm effect1
#> 1 unspecified interaction effect TRUE    FALSE FALSE          0   0    493    
#>   effect2
#> 1 46     

if (FALSE) {
(anspp <- siena07(myalgorithm, data=mydata, effects=myeff, batch=TRUE))
## anspp contains identical results as ansp above.
}

## A demonstration of RateX heterogeneity. Note that rate interactions are
## not implemented in general, just for Rate x cCovar.
if (FALSE) {
myalgorithm <- sienaAlgorithmCreate(nsub=4, n3=1000)
mynet1 <- sienaDependent(array(c(s501, s502, s503), dim=c(50, 50, 3)))
myccov <- coCovar(s50a[,1])
mydata <- sienaDataCreate(mynet1, myccov)
myeff <- getEffects(mydata)
myeff <- includeEffects(myeff, transTrip, balance)
myeff <- includeTimeDummy(myeff, RateX, type="rate", interaction1="myccov")
ans <- siena07(myalgorithm, data=mydata, effects=myeff, batch=TRUE)
}