Versions of FRAN

The functions to be called as "FRAN" by siena07. They call compiled C++.

simstats0c(z, x, data=NULL, effects=NULL, fromFiniteDiff=FALSE,
           returnDeps=FALSE, returnChains=FALSE, byWave=FALSE,
           returnDataFrame=FALSE, returnLoglik=FALSE)
maxlikec(z, x, data=NULL, effects=NULL,
           returnChains=FALSE, byGroup = FALSE, byWave=FALSE,
           returnDataFrame=FALSE, returnLoglik=FALSE,
           onlyLoglik=FALSE)
initializeFRAN(z, x, data, effects, prevAns = NULL, initC,
           profileData = FALSE, returnDeps = FALSE, returnChains =
           FALSE, byGroup = FALSE, returnDataFrame = FALSE,
           byWave = FALSE, returnLoglik = FALSE,  onlyLoglik = FALSE)
terminateFRAN(z, x)

Arguments

z
Control object, passed in automatically in siena07.

x
A sienaAlgorithm object, passed in automatically in siena07.

data
A sienaData object as returned by sienaDataCreate.

effects
A sienaEffects object as returned by getEffects.

fromFiniteDiff
Boolean used during calculation of derivatives by finite differences. Not for user use.

returnDeps
Boolean. Whether to return the simulated networks in Phase 3.

returnChains
Boolean. Whether to return the chains.

byWave
Boolean. Whether to return the finite difference or maximum likelihood derivatives by wave (uses a great deal of memory). Only necessary for sienaTimeTest

byGroup
Boolean. For internal use: allows different thetas for each group to be used in sienaBayes.

returnDataFrame
Boolean. Whether to return the chains as lists or data frames.

returnLoglik
Boolean. Whether to return the log likelihood of the simulated chain.

onlyLoglik
Boolean: whether to return just the likelihood for the simulated chain, plus details of steps accepted and rejected.

prevAns
An object of class "sienaFit" as returned by siena07, from which scaling information (derivative matrix and standard deviation of the deviations) will be extracted along with the latest version of the parameters which will be used as the initial values, unless the model requests the use of standard initial values. If the previous model is exactly the same as the current one, Phase 1 will be omitted. If not, any parameter estimates for effects which are included in the new model will be used as initial values, but phase 1 will still be carried out. If the results used as prevAns are a reasonable starting point, this will increase the efficiency of the algorithm.

initC
If TRUE, call is to setup the data and model in C++. For use with multiple processes only.

profileData
Boolean to force dumping of the data for profiling with sienaProfile.exe.

Details

Not for general users' use.
The name of simstats0c or maxlikec should be used for the element FRAN of the model object, the former when using estimation by forward simulation, the latter for maximum likelihood estimation. The arguments with no defaults must be passed in on the call to siena07. initializeFRAN and terminateFRAN are called in both cases.

Value

simstats0c returns a list containing:

fra

Simulated statistics.

sc

Scores with which to calculate the derivative (not phase 2 or if using finite differences or maximum likelihood).

dff

Contributions to the derivative if finite differences

ntim

For conditional processing, time taken.

feasible

Currently set to TRUE.

OK

Could be set to FALSE if serious error has occurred.

sims

A list of simulation results, one for each period. Each list consists of a list for each data object, each of which consists of a list for each network, each of which consists of a list for each period, each component of which is an edgelist in matrix form (the columns are from, to, value) (or vector for behavior variables). Only if returnDeps is TRUE.

maxlikec returns a list containing:
fra

Simulated scores.

dff

Simulated Hessians: stored as lower triangular matrices

ntim

NULL, compatibility only

feasible

Currently set to TRUE.

OK

Could be set to FALSE if serious error has occurred.

dff

Simulated Hessian

sims

NULL, for compatibility only

chain

A list of sampled chains, one for each period. Each list consists of a list for each data object, each of which consists of a list for each network, each of which consists of a list for each period, each component of which is a list or a data frame depending on the value of returnDataFrame. Only if returnChainss is TRUE.

accepts

Number of accepted MH steps by dependent variable (permute steps are counted under first dependent variable)

rejects

Number of rejected MH steps by dependent variable (permute steps are counted under first dependent variable)

aborts

Number of aborted MH steps counted under first dependent variable.

loglik

Loglikelihood of the simulations. Only if returnLoglik is TRUE. If onlyLoglik is TRUE, only loglik, accepts,rejects and aborts are returned.

initializeFRAN and terminateFRAN return the control object z.
References

See http://www.stats.ox.ac.uk/~snijders/siena/

See also

Examples

mynet1 <- sienaNet(array(c(tmp3, tmp4), dim=c(32, 32, 2)))
mydata <- sienaDataCreate(mynet1)
myeff <- getEffects(mydata)
myeff <- includeEffects(myeff, transTrip)

#>   effectName          include fix   test  initialValue parm
#> 1 transitive triplets TRUE    FALSE FALSE          0   0   

myalgorithm <- sienaAlgorithmCreate(fn=simstats0c, nsub=2, n3=100, projname=NULL)

#> If you use this algorithm object, siena07 will create an output file /var/folders/pl/ngry9kn13p7466xwk0_2xyh40000gn/T//RtmpBuM5HN/Siena344e4f774aef.txt .
#> This is a temporary file for this R session.

ans <- siena07(myalgorithm, data=mydata, effects=myeff,  batch=TRUE)

#> 
#> Start phase 0 
#> theta: -0.56  0.00  0.00 
#> 
#> Start phase 1 
#> Phase 1 Iteration 1 Progress: 0%
#> Phase 1 Iteration 2 Progress: 0%
#> Phase 1 Iteration 3 Progress: 0%
#> Phase 1 Iteration 4 Progress: 1%
#> Phase 1 Iteration 5 Progress: 1%
#> Phase 1 Iteration 10 Progress: 1%
#> Phase 1 Iteration 15 Progress: 2%
#> Phase 1 Iteration 20 Progress: 3%
#> Phase 1 Iteration 25 Progress: 3%
#> Phase 1 Iteration 30 Progress: 4%
#> Phase 1 Iteration 35 Progress: 4%
#> Phase 1 Iteration 40 Progress: 5%
#> Phase 1 Iteration 45 Progress: 6%
#> Phase 1 Iteration 50 Progress: 6%
#> theta: -0.7918  0.4229  0.0718 
#> 
#> Start phase 2.1
#> Phase 2 Subphase 1 Iteration 1 Progress: 26%
#> Phase 2 Subphase 1 Iteration 2 Progress: 26%
#> theta -0.954  0.757  0.135 
#> ac 0.291 0.953 2.316 
#> Phase 2 Subphase 1 Iteration 3 Progress: 26%
#> Phase 2 Subphase 1 Iteration 4 Progress: 26%
#> theta -1.44  1.64  0.27 
#> ac 0.850 0.455 0.559 
#> Phase 2 Subphase 1 Iteration 5 Progress: 26%
#> Phase 2 Subphase 1 Iteration 6 Progress: 26%
#> theta -1.705  1.690  0.303 
#> ac 0.836 0.455 0.554 
#> Phase 2 Subphase 1 Iteration 7 Progress: 26%
#> Phase 2 Subphase 1 Iteration 8 Progress: 26%
#> theta -1.899  1.715  0.333 
#> ac 0.846 0.459 0.685 
#> Phase 2 Subphase 1 Iteration 9 Progress: 27%
#> Phase 2 Subphase 1 Iteration 10 Progress: 27%
#> theta -1.940  1.632  0.346 
#> ac 0.880 0.379 0.724 
#> theta -1.777  1.351  0.329 
#> ac -0.00536 -0.15538 -0.21659 
#> theta: -1.777  1.351  0.329 
#> 
#> Start phase 2.2
#> Phase 2 Subphase 2 Iteration 1 Progress: 54%
#> Phase 2 Subphase 2 Iteration 2 Progress: 54%
#> Phase 2 Subphase 2 Iteration 3 Progress: 54%
#> Phase 2 Subphase 2 Iteration 4 Progress: 54%
#> Phase 2 Subphase 2 Iteration 5 Progress: 55%
#> Phase 2 Subphase 2 Iteration 6 Progress: 55%
#> Phase 2 Subphase 2 Iteration 7 Progress: 55%
#> Phase 2 Subphase 2 Iteration 8 Progress: 55%
#> Phase 2 Subphase 2 Iteration 9 Progress: 55%
#> Phase 2 Subphase 2 Iteration 10 Progress: 55%
#> theta -1.77  1.37  0.32 
#> ac -0.000234 -0.290328 -0.011235 
#> theta: -1.77  1.37  0.32 
#> 
#> Start phase 3