Print method for Wald and score tests for RSiena results

This method prints Wald-type and score-type tests for results estimated by siena07.

# S3 method for sienaTest
print(x, ...)

Arguments

x
An object of type sienaTest, produced by Wald.RSiena, Multipar.RSiena, or score.Test.

...
Extra arguments (not used at present).

Details

The functions Wald.RSiena, Multipar.RSiena, and score.Test produce an object of type sienaTest. These can be printed by this method.

Value

An object of type sienaTest.

See also

Examples

mynet <- sienaDependent(array(c(s501, s502), dim=c(50, 50, 2)))
mydata <- sienaDataCreate(mynet)
myeff <- getEffects(mydata)
myalgorithm <- sienaAlgorithmCreate(nsub=1, n3=40, seed=123, projname=NULL)

#> If you use this algorithm object, siena07 will create an output file /var/folders/pl/ngry9kn13p7466xwk0_2xyh40000gn/T//RtmpNeONkK/Siena3ebf6939023c.txt .
#> This is a temporary file for this R session.

# nsub=1 and n3=40 is used here for having a brief computation,
# not for practice.
myeff <- includeEffects(myeff, transTrip, transTies)

#>   effectName          include fix   test  initialValue parm
#> 1 transitive triplets TRUE    FALSE FALSE          0   0   
#> 2 transitive ties     TRUE    FALSE FALSE          0   0   

myeff <- includeEffects(myeff, outAct, outPop, fix=TRUE, test=TRUE)

#>   effectName             include fix  test initialValue parm
#> 1 outdegree - popularity TRUE    TRUE TRUE          0   0   
#> 2 outdegree - activity   TRUE    TRUE TRUE          0   0   

(ans <- siena07(myalgorithm, data=mydata, effects=myeff, batch=TRUE))

#> 
#> Start phase 0 
#> theta: -1.49  0.00  0.00  0.00  0.00  0.00 
#> 
#> Start phase 1 
#> Phase 1 Iteration 1 Progress: 0%
#> Phase 1 Iteration 2 Progress: 0%
#> Phase 1 Iteration 3 Progress: 0%
#> Phase 1 Iteration 4 Progress: 1%
#> Phase 1 Iteration 5 Progress: 1%
#> Phase 1 Iteration 10 Progress: 2%
#> Phase 1 Iteration 15 Progress: 2%
#> Phase 1 Iteration 20 Progress: 3%
#> Phase 1 Iteration 25 Progress: 4%
#> Phase 1 Iteration 30 Progress: 5%
#> Phase 1 Iteration 35 Progress: 6%
#> Phase 1 Iteration 40 Progress: 6%
#> Phase 1 Iteration 45 Progress: 7%
#> Phase 1 Iteration 50 Progress: 8%
#> theta: -1.691  0.394  0.419 -0.246  0.000  0.000 
#> 
#> Start phase 2.1
#> Phase 2 Subphase 1 Iteration 1 Progress: 56%
#> Phase 2 Subphase 1 Iteration 2 Progress: 57%
#> theta -1.813  0.601  0.621 -0.194  0.000  0.000 
#> ac 0.1331 1.8626 5.8043 6.1750 0.0164 0.2778 
#> Phase 2 Subphase 1 Iteration 3 Progress: 57%
#> Phase 2 Subphase 1 Iteration 4 Progress: 57%
#> theta -2.2986  1.6068  0.9898 -0.0947  0.0000  0.0000 
#> ac  0.3221  0.1176 -0.0470 -0.0484  0.2892  0.3335 
#> Phase 2 Subphase 1 Iteration 5 Progress: 57%
#> Phase 2 Subphase 1 Iteration 6 Progress: 57%
#> theta -2.483  2.122  0.619  0.267  0.000  0.000 
#> ac 0.571 0.413 0.169 0.277 0.611 0.591 
#> Phase 2 Subphase 1 Iteration 7 Progress: 57%
#> Phase 2 Subphase 1 Iteration 8 Progress: 58%
#> theta -2.606  2.277  0.263  0.412  0.000  0.000 
#> ac 0.567 0.427 0.151 0.199 0.613 0.594 
#> Phase 2 Subphase 1 Iteration 9 Progress: 58%
#> Phase 2 Subphase 1 Iteration 10 Progress: 58%
#> theta -2.751  2.306  0.273  0.537  0.000  0.000 
#> ac 0.556 0.441 0.145 0.194 0.614 0.607 
#> theta -2.673  2.086  0.291  0.683  0.000  0.000 
#> ac -0.00724 -0.23041 -0.15029 -0.29298  0.23609  0.23514 
#> theta: -2.673  2.086  0.291  0.683  0.000  0.000 
#> 
#> Start phase 3 

#> Estimates, standard errors and convergence t-ratios
#> 
#>                                  Estimate   Standard   Convergence 
#>                                               Error      t-ratio   
#> 
#> Rate parameters: 
#>   0       Rate parameter          6.8214  ( 1.2733   )             
#> 
#> Other parameters: 
#>   1. eval outdegree (density)    -2.6729  ( 0.1762   )   -0.0968   
#>   2. eval reciprocity             2.0859  ( 0.3172   )   -0.1183   
#>   3. eval transitive triplets     0.2913  ( 0.6362   )   -0.0741   
#>   4. eval transitive ties         0.6835  ( 1.3012   )   -0.0611   
#>   5. eval outdegree - popularity  0.0000  (     NA   )    0.6060   
#>   6. eval outdegree - activity    0.0000  (     NA   )    0.4845   
#> 
#> Overall maximum convergence ratio:    0.1743 
#> 
#> 
#> Score test for 2 parameters:
#> chi-squared = 11.49, p = 0.0032.
#> 
#> Total of 166 iteration steps.
#> 

mprs <- Multipar.RSiena(ans, 3, 4)
print(mprs)

#> Tested effects:
#>  mynet: transitive triplets 
#>  mynet: transitive ties 
#> chi-squared = 16.52, d.f. = 2;  p < 0.001.