Methods for processing sienaFit objects, produced by siena07.

print, summary, and xtable methods for sienaFit objects.

# S3 method for sienaFit
print(x, tstat=TRUE, ...)

# S3 method for sienaFit
summary(object, ...)

# S3 method for summary.sienaFit
print(x, matrices=TRUE, ...)

# S3 method for sienaFit
xtable(x, caption = NULL, label = NULL, align = NULL,
                digits = NULL, display = NULL, ...)

siena.table(x, type="tex", file=paste(deparse(substitute(x)), ".", type,sep=""),
            vertLine=TRUE, tstatPrint=FALSE, sig=FALSE, d=3, nfirst=NULL)

Arguments

object
An object of class sienaFit, produced by siena07.

x
An object of class sienaFit, or summary.sienaFit as appropriate.

matrices
Boolean: whether also to print in the summary the covariance matrix of the estimates, the derivative matrix of expected statistics X by parameters, and the covariance matrix of the statistics.

tstat
Boolean: if this is NULL, the t-statistics for convergence will not be added to the report.

type
Type of output to produce; must be either "tex" or "html".

file
Name of the file; defaults to the name of the sienaFit object.

vertLine
Boolean: add vertical lines separating the columns in siena.table.

tstatPrint
Boolean: add a column of significance t values (parameter estimate/standard error estimate) to siena.table.

sig
Boolean: adds symbols (daggers and asterisks) indicating significance levels for the parameter estimates to siena.table.

d
The number of decimals places used in siena.table.

caption
See documentation for xtable.

label
See documentation for xtable.

align
See documentation for xtable.

digits
See documentation for xtable.

display
See documentation for xtable

nfirst
Only relevant for the RSienaTest package.

...
Add extra parameters for print.xtable here. e.g. type, file.

Value

The function print.sienaFit prints a table containing estimated parameter values, standard errors and (optionally) t-statistics for convergence.

The function summary.sienaFit prints a table containing estimated parameter values, standard errors and t-statistics for convergence together with the covariance matrix of the estimates, the derivative matrix of expected statistics X by parameters, and the covariance matrix of the expected statistics X.

The function xtable.sienaFit creates an object of class xtable.sienaFit which inherits from class xtable and passes an extra arguments to the print.xtable.

The function siena.table outputs a latex or html table of the estimates and standards errors of a sienaFit object. The table will be written to a file in the current directory and has a footnote reporting the maximum of the convergence t-ratios.

See the manual for how to import these tables easily into MS-Word.

References

See http://www.stats.ox.ac.uk/~snijders/siena/

See also

Examples

myalgorithm <- sienaAlgorithmCreate(nsub=2, n3=100, projname=NULL)

#> If you use this algorithm object, siena07 will create an output file /var/folders/pl/ngry9kn13p7466xwk0_2xyh40000gn/T//RtmpXe2fRo/Siena6ef923a85cb.txt .
#> This is a temporary file for this R session.

mynet1 <- sienaDependent(array(c(tmp3, tmp4), dim=c(32, 32, 2)))
mydata <- sienaDataCreate(mynet1)
myeff <- getEffects(mydata)
ans <- siena07(myalgorithm, data=mydata, effects=myeff, batch=TRUE)

#> 
#> Start phase 0 
#> theta: -0.56  0.00 
#> 
#> Start phase 1 
#> Phase 1 Iteration 1 Progress: 0%
#> Phase 1 Iteration 2 Progress: 0%
#> Phase 1 Iteration 3 Progress: 0%
#> Phase 1 Iteration 4 Progress: 1%
#> Phase 1 Iteration 5 Progress: 1%
#> Phase 1 Iteration 10 Progress: 1%
#> Phase 1 Iteration 15 Progress: 2%
#> Phase 1 Iteration 20 Progress: 3%
#> Phase 1 Iteration 25 Progress: 3%
#> Phase 1 Iteration 30 Progress: 4%
#> Phase 1 Iteration 35 Progress: 5%
#> Phase 1 Iteration 40 Progress: 6%
#> Phase 1 Iteration 45 Progress: 6%
#> Phase 1 Iteration 50 Progress: 7%
#> theta: -0.646  0.358 
#> 
#> Start phase 2.1
#> Phase 2 Subphase 1 Iteration 1 Progress: 21%
#> Phase 2 Subphase 1 Iteration 2 Progress: 21%
#> theta -0.743  0.836 
#> ac 0.853 1.747 
#> Phase 2 Subphase 1 Iteration 3 Progress: 21%
#> Phase 2 Subphase 1 Iteration 4 Progress: 21%
#> theta -0.943  1.779 
#> ac 0.229 0.871 
#> Phase 2 Subphase 1 Iteration 5 Progress: 21%
#> Phase 2 Subphase 1 Iteration 6 Progress: 21%
#> theta -1.07  2.09 
#> ac 0.497 0.748 
#> Phase 2 Subphase 1 Iteration 7 Progress: 22%
#> Phase 2 Subphase 1 Iteration 8 Progress: 22%
#> theta -1.14  2.08 
#> ac 0.575 0.655 
#> Phase 2 Subphase 1 Iteration 9 Progress: 22%
#> Phase 2 Subphase 1 Iteration 10 Progress: 22%
#> theta -1.19  1.95 
#> ac 0.573 0.615 
#> theta -1.13  1.77 
#> ac 0.0869 0.1281 
#> theta: -1.13  1.77 
#> 
#> Start phase 2.2
#> Phase 2 Subphase 2 Iteration 1 Progress: 51%
#> Phase 2 Subphase 2 Iteration 2 Progress: 51%
#> Phase 2 Subphase 2 Iteration 3 Progress: 52%
#> Phase 2 Subphase 2 Iteration 4 Progress: 52%
#> Phase 2 Subphase 2 Iteration 5 Progress: 52%
#> Phase 2 Subphase 2 Iteration 6 Progress: 52%
#> Phase 2 Subphase 2 Iteration 7 Progress: 52%
#> Phase 2 Subphase 2 Iteration 8 Progress: 52%
#> Phase 2 Subphase 2 Iteration 9 Progress: 52%
#> Phase 2 Subphase 2 Iteration 10 Progress: 53%
#> theta -1.15  1.78 
#> ac 0.0131 0.1368 
#> theta: -1.15  1.78 
#> 
#> Start phase 3 

ans

#> Estimates, standard errors and convergence t-ratios
#> 
#>                               Estimate   Standard   Convergence 
#>                                            Error      t-ratio   
#> 
#> Rate parameters: 
#>   0       Rate parameter       3.0656  ( 0.4372   )             
#> 
#> Other parameters: 
#>   1. eval outdegree (density) -1.1461  ( 0.2070   )   -0.0055   
#>   2. eval reciprocity          1.7805  ( 0.4551   )    0.0202   
#> 
#> Overall maximum convergence ratio:    0.0266 
#> 
#> 
#> Total of 627 iteration steps.
#> 

summary(ans)

#> Estimates, standard errors and convergence t-ratios
#> 
#>                               Estimate   Standard   Convergence 
#>                                            Error      t-ratio   
#> 
#> Rate parameters: 
#>   0       Rate parameter       3.0656  ( 0.4372   )             
#> 
#> Other parameters: 
#>   1. eval outdegree (density) -1.1461  ( 0.2070   )   -0.0055   
#>   2. eval reciprocity          1.7805  ( 0.4551   )    0.0202   
#> 
#> Overall maximum convergence ratio:    0.0266 
#> 
#> 
#> Total of 627 iteration steps.
#> 
#> Covariance matrix of estimates (correlations below diagonal)
#> 
#>        0.043       -0.057
#>       -0.602        0.207
#> 
#> Derivative matrix of expected statistics X by parameters:
#> 
#>       44.238       17.097
#>       12.259       16.187
#> 
#> Covariance matrix of X (correlations below diagonal):
#> 
#>       53.413       20.975
#>        0.482       35.379

if (FALSE) {
xtable(ans, type="html", file="ans.html")
siena.table(ans, type="html", tstat=TRUE, d=2)
}