A function to update the initial values of theta, and a function to update an effects object.

updateTheta copies the final values of any matching selected effects from a sienaFit object to a Siena effects object.
updateSpecification includes in a Siena effects object a set of effects that are included in another effects object.

updateTheta(effects, prevAns, varName=NULL)
updateSpecification(effects.to, effects.from, name.to=NULL, name.from=NULL)

Arguments

effects
Object of class sienaEffects.

prevAns
Object of class sienaFit as returned by siena07.

varName
Character string or vector of character strings; is this is not NULL, the update will only applied to this dependent variable / these dependent variables.

effects.to
Object of class sienaEffects.

effects.from
Object of class sienaEffects.

name.to
Character string, name of dependent variable in object.to.

name.from
Character string, name of dependent variable in object.from.

Details

The initial values of any selected effects in the input effects object which match an effect estimated in prevAns will be updated by updateTheta. If the previous run was conditional, the estimated rate parameters for the dependent variable on which the run was conditioned are added to the final value of theta. If varName is not NULL, this update is restricted to effects for the dependent variable/s specified by varName.
By updateSpecification, the effects included in effects.from are also included in effects.to; if name.to and/or name.from is specified, this is restricted to effects for those dependent variables. If name.to = "all" (should then not be used as variable name!), the effects for all dependent variables will be updated.
Correspondence between effects is defined by "name", "shortName" "type", "groupName", "interaction1", "interaction2", "period", "effect1", "effect2", and "effect3". This means that inclusion of user-defined interactions will be updated only if they were available (i.e., defined) already in effects.to.

Value

Updated effects object.

Note

Using updateTheta explicitly before calling siena07 rather than using it via the argument prevAns of siena07 will not permit the use of the previous derivative matrix. In most cases, using siena07 with prevAns will be more efficient.

References

See http://www.stats.ox.ac.uk/~snijders/siena/

See also

Examples

## For updateTheta:
mynet1 <- sienaDependent(array(c(tmp3, tmp4), dim=c(32, 32, 2)))
mydata <- sienaDataCreate(mynet1)
myeff1 <- getEffects(mydata)
myeff1 <- includeEffects(myeff1, transTrip)

#>   effectName          include fix   test  initialValue parm
#> 1 transitive triplets TRUE    FALSE FALSE          0   0   

myalgorithm <- sienaAlgorithmCreate(nsub=1, n3=100, projname=NULL)

#> If you use this algorithm object, siena07 will create an output file /var/folders/pl/ngry9kn13p7466xwk0_2xyh40000gn/T//RtmpNeONkK/Siena3ebffc2d049.txt .
#> This is a temporary file for this R session.

ans <- siena07(myalgorithm, data=mydata, effects=myeff1, batch=TRUE)

#> 
#> Start phase 0 
#> theta: -0.56  0.00  0.00 
#> 
#> Start phase 1 
#> Phase 1 Iteration 1 Progress: 0%
#> Phase 1 Iteration 2 Progress: 0%
#> Phase 1 Iteration 3 Progress: 1%
#> Phase 1 Iteration 4 Progress: 1%
#> Phase 1 Iteration 5 Progress: 1%
#> Phase 1 Iteration 10 Progress: 2%
#> Phase 1 Iteration 15 Progress: 3%
#> Phase 1 Iteration 20 Progress: 4%
#> Phase 1 Iteration 25 Progress: 5%
#> Phase 1 Iteration 30 Progress: 6%
#> Phase 1 Iteration 35 Progress: 7%
#> Phase 1 Iteration 40 Progress: 8%
#> Phase 1 Iteration 45 Progress: 9%
#> Phase 1 Iteration 50 Progress: 10%
#> theta: -0.7761  0.1134  0.0571 
#> 
#> Start phase 2.1
#> Phase 2 Subphase 1 Iteration 1 Progress: 38%
#> Phase 2 Subphase 1 Iteration 2 Progress: 38%
#> theta -0.944  0.347  0.104 
#> ac -0.265  1.910  1.215 
#> Phase 2 Subphase 1 Iteration 3 Progress: 39%
#> Phase 2 Subphase 1 Iteration 4 Progress: 39%
#> theta -1.476  0.865  0.253 
#> ac -0.123  2.117  1.332 
#> Phase 2 Subphase 1 Iteration 5 Progress: 39%
#> Phase 2 Subphase 1 Iteration 6 Progress: 39%
#> theta -1.759  1.172  0.346 
#> ac -0.168  1.262  1.128 
#> Phase 2 Subphase 1 Iteration 7 Progress: 39%
#> Phase 2 Subphase 1 Iteration 8 Progress: 40%
#> theta -1.883  1.233  0.375 
#> ac 0.267 1.278 1.067 
#> Phase 2 Subphase 1 Iteration 9 Progress: 40%
#> Phase 2 Subphase 1 Iteration 10 Progress: 40%
#> theta -2.023  1.392  0.397 
#> ac 0.287 1.275 0.972 
#> theta -1.749  1.298  0.327 
#> ac -0.18045 -0.02041 -0.00345 
#> theta: -1.749  1.298  0.327 
#> 
#> Start phase 3 

ans$theta

#> [1] -1.7487592  1.2976930  0.3272358

(myeff <- updateTheta(myeff1, ans))

#>   effectName                  include fix   test  initialValue parm
#> 1 basic rate parameter mynet1 TRUE    FALSE FALSE    3.08688   0   
#> 2 outdegree (density)         TRUE    FALSE FALSE   -1.74876   0   
#> 3 reciprocity                 TRUE    FALSE FALSE    1.29769   0   
#> 4 transitive triplets         TRUE    FALSE FALSE    0.32724   0   

##
## For updateSpecification:
myeff2 <- getEffects(mydata)
myeff2 <- includeEffects(myeff2, inPop)

#>   effectName            include fix   test  initialValue parm
#> 1 indegree - popularity TRUE    FALSE FALSE          0   0   

updateSpecification(myeff2, myeff1)

#>   effectName                  include fix   test  initialValue parm
#> 1 basic rate parameter mynet1 TRUE    FALSE FALSE    4.80941   0   
#> 2 outdegree (density)         TRUE    FALSE FALSE   -0.56039   0   
#> 3 reciprocity                 TRUE    FALSE FALSE    0.00000   0   
#> 4 transitive triplets         TRUE    FALSE FALSE    0.00000   0   
#> 5 indegree - popularity       TRUE    FALSE FALSE    0.00000   0   

# Create (meaningless) two-dimensional dependent network
mynet1 <- sienaDependent(array(c(s501, s502), dim=c(50, 50, 2)))
mynet2 <- sienaDependent(array(c(s503, s501), dim=c(50, 50, 2)))
mydata12 <- sienaDataCreate(mynet1, mynet2)
myeff12 <- getEffects(mydata12)
myeff.new <- getEffects(mydata12)
(myeff12 <- includeEffects(myeff12, inPop, outPop, outAct))

#>   effectName                     include fix   test  initialValue parm
#> 1 mynet1: indegree - popularity  TRUE    FALSE FALSE          0   0   
#> 2 mynet1: outdegree - popularity TRUE    FALSE FALSE          0   0   
#> 3 mynet1: outdegree - activity   TRUE    FALSE FALSE          0   0   

#>   name   effectName                     include fix   test  initialValue parm
#> 1 mynet1 basic rate parameter mynet1    TRUE    FALSE FALSE    4.69604   0   
#> 2 mynet1 mynet1: outdegree (density)    TRUE    FALSE FALSE   -1.48852   0   
#> 3 mynet1 mynet1: reciprocity            TRUE    FALSE FALSE    0.00000   0   
#> 4 mynet1 mynet1: indegree - popularity  TRUE    FALSE FALSE    0.00000   0   
#> 5 mynet1 mynet1: outdegree - popularity TRUE    FALSE FALSE    0.00000   0   
#> 6 mynet1 mynet1: outdegree - activity   TRUE    FALSE FALSE    0.00000   0   
#> 7 mynet2 basic rate parameter mynet2    TRUE    FALSE FALSE    5.43044   0   
#> 8 mynet2 mynet2: outdegree (density)    TRUE    FALSE FALSE   -1.54214   0   
#> 9 mynet2 mynet2: reciprocity            TRUE    FALSE FALSE    0.00000   0   

# update myeff.new only for mynet1:
updateSpecification(myeff.new, myeff12)

#>   name   effectName                     include fix   test  initialValue parm
#> 1 mynet1 basic rate parameter mynet1    TRUE    FALSE FALSE    4.69604   0   
#> 2 mynet1 mynet1: outdegree (density)    TRUE    FALSE FALSE   -1.48852   0   
#> 3 mynet1 mynet1: reciprocity            TRUE    FALSE FALSE    0.00000   0   
#> 4 mynet1 mynet1: indegree - popularity  TRUE    FALSE FALSE    0.00000   0   
#> 5 mynet1 mynet1: outdegree - popularity TRUE    FALSE FALSE    0.00000   0   
#> 6 mynet1 mynet1: outdegree - activity   TRUE    FALSE FALSE    0.00000   0   
#> 7 mynet2 basic rate parameter mynet2    TRUE    FALSE FALSE    5.43044   0   
#> 8 mynet2 mynet2: outdegree (density)    TRUE    FALSE FALSE   -1.54214   0   
#> 9 mynet2 mynet2: reciprocity            TRUE    FALSE FALSE    0.00000   0   

# update myeff.new for all dependent networks:
(myeff.updated <- updateSpecification(myeff.new, myeff12, "all"))

#>    name   effectName                     include fix   test  initialValue parm
#> 1  mynet1 basic rate parameter mynet1    TRUE    FALSE FALSE    4.69604   0   
#> 2  mynet1 mynet1: outdegree (density)    TRUE    FALSE FALSE   -1.48852   0   
#> 3  mynet1 mynet1: reciprocity            TRUE    FALSE FALSE    0.00000   0   
#> 4  mynet1 mynet1: indegree - popularity  TRUE    FALSE FALSE    0.00000   0   
#> 5  mynet1 mynet1: outdegree - popularity TRUE    FALSE FALSE    0.00000   0   
#> 6  mynet1 mynet1: outdegree - activity   TRUE    FALSE FALSE    0.00000   0   
#> 7  mynet2 basic rate parameter mynet2    TRUE    FALSE FALSE    5.43044   0   
#> 8  mynet2 mynet2: outdegree (density)    TRUE    FALSE FALSE   -1.54214   0   
#> 9  mynet2 mynet2: reciprocity            TRUE    FALSE FALSE    0.00000   0   
#> 10 mynet2 mynet2: indegree - popularity  TRUE    FALSE FALSE    0.00000   0   
#> 11 mynet2 mynet2: outdegree - popularity TRUE    FALSE FALSE    0.00000   0   
#> 12 mynet2 mynet2: outdegree - activity   TRUE    FALSE FALSE    0.00000   0   

# use multivariate effects object to update univariate effects object:
myeff1 <- getEffects(sienaDataCreate(mynet1))
updateSpecification(myeff1, myeff.updated)

#>   effectName                  include fix   test  initialValue parm
#> 1 basic rate parameter mynet1 TRUE    FALSE FALSE    4.69604   0   
#> 2 outdegree (density)         TRUE    FALSE FALSE   -1.48852   0   
#> 3 reciprocity                 TRUE    FALSE FALSE    0.00000   0   
#> 4 indegree - popularity       TRUE    FALSE FALSE    0.00000   0   
#> 5 outdegree - popularity      TRUE    FALSE FALSE    0.00000   0   
#> 6 outdegree - activity        TRUE    FALSE FALSE    0.00000   0